Refactor: Babel - Split imported static evaluation
