+++ packages/babel/src/staticCssEval/cjsRequireAnalysis.ts
@@ -188,14 +188,12 @@ function unwrapTransparentExpression(expression: t.Expression): t.Expression {
   return currentExpression;
 }
 
-// == Tests ====================================================================
-// Ignore errors when compiling to CommonJS.
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore error TS1343: The 'import.meta' meta-property is only allowed when the '--module' option is 'es2020', 'es2022', 'esnext', 'system', 'node16', or 'nodenext'.
-if (import.meta.vitest) {
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore error TS1343: The 'import.meta' meta-property is only allowed when the '--module' option is 'es2020', 'es2022', 'esnext', 'system', 'node16', or 'nodenext'.
-  const { describe, expect, it } = import.meta.vitest;
+const vitest = import.meta.vitest;
+
+if (vitest) {
+  const { describe, expect, it } = vitest;
 
   function analyzeInitializer(source: string, bindingName: string): boolean {
     let result: boolean | null = null;
