+++ packages/babel/src/staticCssEval/cjsResolver.ts
@@ -73,14 +73,12 @@ function createStaticCssEvalCjsRequestParts(options: {
   };
 }
 
-// == Tests ====================================================================
-// Ignore errors when compiling to CommonJS.
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore error TS1343: The 'import.meta' meta-property is only allowed when the '--module' option is 'es2020', 'es2022', 'esnext', 'system', 'node16', or 'nodenext'.
-if (import.meta.vitest) {
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore error TS1343: The 'import.meta' meta-property is only allowed when the '--module' option is 'es2020', 'es2022', 'esnext', 'system', 'node16', or 'nodenext'.
-  const { describe, expect, it } = import.meta.vitest;
+const vitest = import.meta.vitest;
+
+if (vitest) {
+  const { describe, expect, it } = vitest;
 
   describe("CommonJS resolution request parts", () => {
     it("maps direct member and destructured require sources through one request shape", () => {
