+++ packages/babel/src/staticCssEval/cycles.ts
@@ -151,10 +151,10 @@ function getStaticCssEvalCycleDependencies(
 // Ignore errors when compiling to CommonJS.
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore error TS1343: The 'import.meta' meta-property is only allowed when the '--module' option is 'es2020', 'es2022', 'esnext', 'system', 'node16', or 'nodenext'.
-if (import.meta.vitest) {
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore error TS1343: The 'import.meta' meta-property is only allowed when the '--module' option is 'es2020', 'es2022', 'esnext', 'system', 'node16', or 'nodenext'.
-  const { describe, expect, it } = import.meta.vitest;
+const vitest = import.meta.vitest;
+
+if (vitest) {
+  const { describe, expect, it } = vitest;
 
   const owner = { file: "/project/src/App.tsx", start: 10, end: 30 };
