+++ packages/babel/src/staticCssEval/importedModules.ts
@@ -1,3953 +1,57 @@
 import { types as t } from "@babel/core";
-import type { NodePath } from "@babel/core";
-import hash from "@emotion/hash";
-import type { Binding } from "@babel/traverse";
-import {
-  getStaticObjectMemberValue,
-  getStaticObjectPropertyName,
-  getUnsupportedLiteralReason
-} from "./ast.js";
-import {
-  createStaticCssEvalCandidate,
-  getStaticCssEvalMemberReference,
-  unwrapTransparentCssRuleExpression
-} from "./candidates.js";
-import type { StaticMemberReference } from "./candidates.js";
-import {
-  createStaticCssEvalProviderSourceMetadata,
-  enforceStaticCssEvalProviderSourcePolicy
-} from "./boundary.js";
-import {
-  collectStaticCssEvalCjsRequireBindings,
-  getStaticCssEvalCjsRequireSource,
-  isStaticCssEvalLiteralRequireCallExpression
-} from "./cjsBindings.js";
-import type {
-  ImportedStaticCssEvalCjsBinding,
-  StaticCssEvalCjsRequireSource
-} from "./cjsBindings.js";
-import { containsStaticCssEvalRequireCallExpression } from "./cjsRequireAnalysis.js";
-import {
-  createStaticCssEvalCjsImportRequestParts,
-  createStaticCssEvalCjsReexportRequestParts
-} from "./cjsResolver.js";
-import type { StaticCssEvalCjsResolutionRequestParts } from "./cjsResolver.js";
-import type {
-  StaticCssEvalProviderSourceMetadata,
-  StaticCssEvalProviderSourcePolicyDescriptor
-} from "./boundary.js";
-import {
-  createStaticCssEvalAmbiguousExportStarDiagnostic,
-  createStaticCssEvalCjsBundleRuntimeUnsupportedDiagnostic,
-  createStaticCssEvalCjsDynamicRequireUnsupportedDiagnostic,
-  createStaticCssEvalCjsExportUnsupportedDiagnostic,
-  createStaticCssEvalCjsHelperUnsupportedDiagnostic,
-  createStaticCssEvalCjsUnsupportedDiagnostic,
-  createStaticCssEvalComputedMemberUnsupportedDiagnostic,
-  createStaticCssEvalDiagnostic,
-  createStaticCssEvalDynamicExpressionUnsupportedDiagnostic,
-  createStaticCssEvalExportStarUnsupportedDiagnostic,
-  createStaticCssEvalMutableBindingDiagnostic,
-  createStaticCssEvalMutatedBindingDiagnostic,
-  createStaticCssEvalNamespaceReexportUnsupportedDiagnostic,
-  createStaticCssEvalObjectSpreadUnsupportedDiagnostic,
-  createStaticCssEvalPartialNamespaceFailureDiagnostic,
-  createStaticCssEvalProviderSourceUnsupportedDiagnostic,
-  createStaticCssEvalUnresolvedExportDiagnostic,
-  createStaticCssEvalUnresolvedImportDiagnostic,
-  guardStaticCssEvalImportCycle,
-  guardStaticCssEvalResolutionDepth
-} from "./diagnostics.js";
-import type { StaticCssEvalImportCycleKey } from "./diagnostics.js";
-import {
-  enforceStaticCssEvalLiteralNodeCount,
-  enforceStaticCssEvalObjectArrayRecursionDepth,
-  enforceStaticCssEvalSourceSize
-} from "./limits.js";
+import { getUnsupportedLiteralReason } from "./ast.js";
 import {
   createExportMapCacheKey,
-  createStaticCssModuleExportNameTable,
   createStaticCssModuleCache,
   formatExportMapCacheKey,
   STATIC_CSS_MODULE_CACHE_PARSER_VERSION,
   STATIC_CSS_MODULE_CACHE_SUPPORT_VERSION
 } from "./moduleCache.js";
+import type { StaticCssModuleCache } from "./moduleCache.js";
 import type {
-  ExportMapEntry,
-  ExportGraphStarReexportEntry,
-  ExportMapLocalEntry,
-  ParsedStaticCssModule,
-  StaticCssModuleCache,
-  StaticCssModuleExportNameTableEntry,
-  StaticCssModuleExportStarSource,
-  StaticCssModuleSource
-} from "./moduleCache.js";
-import {
-  getStaticCssEvalConstBindingInitExpression,
-  hasStaticCssEvalBindingMutation
-} from "./sameFile.js";
-import type {
-  BindingProvenance,
-  ResolutionChainEntry,
   ResolutionDependency,
-  ResolutionDependencyKind,
   StaticCssEvalCacheKey,
-  StaticCssEvalDiagnostic,
   StaticCssEvalDiagnosticId,
-  StaticCssEvalExportName,
   StaticCssEvalProvider,
-  StaticCssEvalQuery,
   StaticCssEvalResult,
-  StaticCssEvalSourceLocation,
-  StaticCssEvalUnsupportedReason,
-  StaticCssEvalModuleRecord,
   StaticCssLiteral
 } from "./types.js";
-
-export type ImportedStaticCssEvalImportBinding =
-  | {
-      kind: "default";
-      localName: string;
-      importPath: string;
-      importedName: "default";
-    }
-  | {
-      kind: "named";
-      localName: string;
-      importPath: string;
-      importedName: string;
-    }
-  | {
-      kind: "namespace";
-      localName: string;
-      importPath: string;
-    };
-
-export type { ImportedStaticCssEvalCjsBinding };
-
-export type ImportedStaticCssEvalExportBinding = ExportMapEntry;
-
-type ImportedStaticCssEvalResolutionResult =
-  | { kind: "not-candidate" }
-  | {
-      kind: "resolved";
-      value: StaticCssLiteral;
-      provenance: BindingProvenance;
-      cacheKey: StaticCssEvalCacheKey;
-    }
-  | {
-      kind: "error";
-      diagnostic: StaticCssEvalDiagnostic;
-      provenance?: BindingProvenance;
-      cacheKey?: StaticCssEvalCacheKey;
-    };
-
-type ImportedStaticCssEvalExportPresenceResult =
-  | { kind: "found" }
-  | { kind: "missing" }
-  | { kind: "error"; diagnostic: StaticCssEvalDiagnostic };
-
-interface ImportedStaticCssEvalExportStarCandidate {
-  readonly entry: ExportGraphStarReexportEntry;
-  readonly record: ImportedStaticCssEvalModuleRecord;
-  readonly request: ImportedStaticCssEvalResolutionRequest;
-}
-
-interface ImportedStaticCssEvalNamespaceExportNameTable {
-  readonly table: ReadonlyMap<
-    StaticCssEvalExportName,
-    StaticCssModuleExportNameTableEntry
-  >;
-  readonly candidates: ImportedStaticCssEvalExportStarCandidate[];
-}
-
-interface ImportedStaticCssEvalContext {
-  owner: StaticCssEvalSourceLocation;
-  dependency: StaticCssEvalSourceLocation;
-  importPath: string;
-  exportName: StaticCssEvalExportName;
-  memberPath: string[];
-}
-
-type ImportedStaticCssEvalProvenanceKind = Exclude<
-  BindingProvenance["kind"],
-  "local"
->;
-
-interface ImportedStaticCssEvalResolutionRequest {
-  importer: string;
-  specifier: string;
-  exportName: StaticCssEvalExportName;
-  memberPath: string[];
-  dependencyKind: ResolutionDependencyKind;
-  provenanceKind: ImportedStaticCssEvalProvenanceKind;
-  namespaceBinding?: string;
-  reexportName?: string;
-  wholeNamespace?: boolean;
-}
-
-interface ImportedStaticCssEvalResolvedImport {
-  readonly resolvedId: string;
-  readonly sourceMetadata: StaticCssEvalProviderSourceMetadata;
-}
-
-interface ImportedStaticCssEvalResolutionState {
-  owner: StaticCssEvalSourceLocation;
-  dependencies: Map<string, ResolutionDependency>;
-  resolutionChain: ResolutionChainEntry[];
-  stack: StaticCssEvalImportCycleKey[];
-}
-
-interface StaticCssLiteralValidationState {
-  count: number;
-}
-
-interface ImportedStaticCssEvalLocalReferenceFrame {
-  readonly recordId: string;
-  readonly bindingName: string;
-  readonly memberPath: readonly string[];
-}
-
-type ImportedStaticCssEvalLiteralResult =
-  | { kind: "resolved"; value: StaticCssLiteral }
-  | { kind: "error"; diagnostic: StaticCssEvalDiagnostic };
-
-type ImportedStaticCssEvalLiteralReferenceResult =
-  | ImportedStaticCssEvalLiteralResult
-  | { kind: "not-candidate" };
-
-type ImportedStaticCssEvalDeclarationKind =
-  | "const"
-  | "let"
-  | "var"
-  | "function"
-  | "class";
-
-interface ImportedStaticCssEvalLiteralEvaluationOptions {
-  readonly expression: t.Expression;
-  readonly context: ImportedStaticCssEvalContext;
-  readonly record: ImportedStaticCssEvalModuleRecord;
-  readonly request: ImportedStaticCssEvalResolutionRequest;
-  readonly resolutionState: ImportedStaticCssEvalResolutionState;
-  readonly literalState: StaticCssLiteralValidationState;
-  readonly localStack: readonly ImportedStaticCssEvalLocalReferenceFrame[];
-  readonly depth: number;
-  readonly resolveReference: ImportedStaticCssEvalLiteralReferenceResolver;
-}
-
-interface ImportedStaticCssEvalLiteralReferenceOptions extends ImportedStaticCssEvalLiteralEvaluationOptions {
-  readonly reference: StaticMemberReference;
-}
-
-type ImportedStaticCssEvalLiteralReferenceResolver = (
-  options: ImportedStaticCssEvalLiteralReferenceOptions
-) => ImportedStaticCssEvalLiteralResult;
-
-export interface ImportedStaticCssEvalLoadedModule extends StaticCssEvalProviderSourcePolicyDescriptor {
-  id: string;
-  source: string;
-  realpath?: string;
-  sourceHash?: string;
-  version?: string | number;
-}
-
-export interface ImportedStaticCssEvalImportResolution extends StaticCssEvalProviderSourcePolicyDescriptor {
-  importerId: string;
-  importPath: string;
-  resolvedId: string;
-}
-
-export interface ImportedStaticCssEvalModuleRecord extends StaticCssEvalModuleRecord {
-  source: string;
-  imports: ReadonlyMap<string, ImportedStaticCssEvalImportBinding>;
-  cjsImports: ReadonlyMap<string, ImportedStaticCssEvalCjsBinding>;
-  exports: ReadonlyMap<
-    StaticCssEvalExportName,
-    ImportedStaticCssEvalExportBinding
-  >;
-  exportAllReexportSources: readonly string[];
-  parsedModule: ParsedStaticCssModule;
-  programPath: NodePath<t.Program>;
-}
-
-export interface CreateImportedStaticCssEvalProviderOptions {
-  modules: readonly ImportedStaticCssEvalLoadedModule[];
-  importResolutions: readonly ImportedStaticCssEvalImportResolution[];
-  moduleRecords?: readonly ImportedStaticCssEvalModuleRecord[];
-}
-
-interface ImportedStaticCssEvalResolverOptions extends CreateImportedStaticCssEvalProviderOptions {
-  moduleCache?: StaticCssModuleCache;
-}
-
-export type ResolveImportedStaticCssEvalExpressionResult =
-  | { kind: "not-candidate" }
-  | { kind: "resolved"; expression: t.ObjectExpression | t.ArrayExpression }
-  | { kind: "error"; diagnostic: StaticCssEvalDiagnostic };
-
-const importResolutionKeySeparator = "\0";
-
-export function createImportedStaticCssEvalProvider(
-  options: CreateImportedStaticCssEvalProviderOptions
-): StaticCssEvalProvider {
-  const resolver = new ImportedStaticCssEvalResolver(options);
-
-  return {
-    getResolvedCssValue(query) {
-      return resolver.resolve(query);
-    }
-  };
-}
-
-export function createImportedStaticCssEvalModuleRecord(
-  loadedModule: ImportedStaticCssEvalLoadedModule
-): ImportedStaticCssEvalModuleRecord {
-  return createImportedStaticCssEvalModuleRecordWithCache(
-    loadedModule,
-    createStaticCssModuleCache()
-  );
-}
-
-function createImportedStaticCssEvalModuleRecordWithCache(
-  loadedModule: ImportedStaticCssEvalLoadedModule,
-  moduleCache: StaticCssModuleCache
-): ImportedStaticCssEvalModuleRecord {
-  const parsedModule = moduleCache.getParsedModule(
-    createStaticCssModuleSource(loadedModule)
-  );
-
-  const imports = new Map<string, ImportedStaticCssEvalImportBinding>();
-  const cjsImports = collectStaticCssEvalCjsRequireBindings(
-    parsedModule.programPath
-  );
-
-  for (const statement of parsedModule.program.body) {
-    collectModuleImportBindings(statement, imports);
-  }
-
-  return {
-    id: loadedModule.id,
-    realpath: loadedModule.realpath ?? loadedModule.id,
-    sourceHash:
-      loadedModule.sourceHash ??
-      createInlineStaticCssSourceHash(loadedModule.source),
-    ...(loadedModule.version !== undefined
-      ? { version: loadedModule.version }
-      : {}),
-    ...createStaticCssEvalProviderSourceMetadata(loadedModule),
-    dependencies: [],
-    source: loadedModule.source,
-    imports,
-    cjsImports,
-    exports: parsedModule.exportMap,
-    exportAllReexportSources: parsedModule.unsupportedExportStars.flatMap(
-      (entry) => (entry.source ? [entry.source] : [])
-    ),
-    parsedModule,
-    programPath: parsedModule.programPath
-  };
-}
-
-export function resolveImportedStaticCssEvalExpression(options: {
-  expression: t.Expression;
-  ownerFile: string;
-  provider?: StaticCssEvalProvider;
-  allowUnsupportedSourceFallback?: boolean;
-}): ResolveImportedStaticCssEvalExpressionResult {
-  if (!options.provider) {
-    return { kind: "not-candidate" };
-  }
-
-  const candidate = createStaticCssEvalCandidate(
-    options.expression,
-    options.ownerFile
-  );
-
-  if (!candidate?.bindingName) {
-    return { kind: "not-candidate" };
-  }
-
-  const result = options.provider.getResolvedCssValue(candidate);
-
-  if (result.kind === "not-candidate") {
-    return { kind: "not-candidate" };
-  }
-
-  if (result.kind === "error") {
-    if (
-      options.allowUnsupportedSourceFallback === true &&
-      result.diagnostic.reason === "reexport-or-barrel" &&
-      result.diagnostic.id === undefined
-    ) {
-      return { kind: "not-candidate" };
-    }
-
-    return { kind: "error", diagnostic: result.diagnostic };
-  }
-
-  if (!isStaticCssRuleLiteralValue(result.value)) {
-    return { kind: "not-candidate" };
-  }
-
-  const expression = createStaticCssLiteralExpression(result.value);
-
-  return t.isObjectExpression(expression) || t.isArrayExpression(expression)
-    ? { kind: "resolved", expression }
-    : { kind: "not-candidate" };
-}
-
-export function findUnsupportedImportedStaticCssEvalReferenceDiagnostic(options: {
-  expression: t.Expression;
-  ownerFile: string;
-  provider?: StaticCssEvalProvider;
-}): StaticCssEvalDiagnostic | null {
-  const { provider } = options;
-
-  if (!provider) {
-    return null;
-  }
-
-  const unwrappedExpression = unwrapTransparentCssRuleExpression(
-    options.expression
-  );
-
-  if (t.isObjectExpression(unwrappedExpression)) {
-    return findUnsupportedImportedObjectReferenceDiagnostic(
-      unwrappedExpression,
-      { ownerFile: options.ownerFile, provider }
-    );
-  }
-
-  if (t.isArrayExpression(unwrappedExpression)) {
-    return findUnsupportedImportedArrayReferenceDiagnostic(
-      unwrappedExpression,
-      { ownerFile: options.ownerFile, provider }
-    );
-  }
-
-  return findUnsupportedImportedExpressionReferenceDiagnostic(
-    unwrappedExpression,
-    {
-      ownerFile: options.ownerFile,
-      provider,
-      includeSupportedReferenceErrors: false
-    }
-  );
-}
-
-export function createStaticCssLiteralExpression(
-  value: StaticCssLiteral
-): t.Expression {
-  if (Array.isArray(value)) {
-    return t.arrayExpression(value.map(createStaticCssLiteralExpression));
-  }
-
-  if (value && typeof value === "object") {
-    return t.objectExpression(
-      Object.entries(value).map(([key, propertyValue]) =>
-        t.objectProperty(
-          t.isValidIdentifier(key) ? t.identifier(key) : t.stringLiteral(key),
-          createStaticCssLiteralExpression(propertyValue)
-        )
-      )
-    );
-  }
-
-  if (typeof value === "string") {
-    return t.stringLiteral(value);
-  }
-
-  if (typeof value === "number") {
-    return t.numericLiteral(value);
-  }
-
-  if (typeof value === "boolean") {
-    return t.booleanLiteral(value);
-  }
-
-  return t.nullLiteral();
-}
-
-class ImportedStaticCssEvalResolver {
-  readonly #modules = new Map<string, ImportedStaticCssEvalLoadedModule>();
-  readonly #records = new Map<string, ImportedStaticCssEvalModuleRecord>();
-  readonly #parseFailures = new Map<string, string>();
-  readonly #importResolutions = new Map<
-    string,
-    ImportedStaticCssEvalResolvedImport
-  >();
-  readonly #moduleCache: StaticCssModuleCache;
-
-  constructor(options: ImportedStaticCssEvalResolverOptions) {
-    this.#moduleCache = options.moduleCache ?? createStaticCssModuleCache();
-
-    for (const loadedModule of options.modules) {
-      this.#modules.set(loadedModule.id, loadedModule);
-    }
-
-    for (const moduleRecord of options.moduleRecords ?? []) {
-      this.#records.set(moduleRecord.id, moduleRecord);
-    }
-
-    for (const resolution of options.importResolutions) {
-      this.#importResolutions.set(
-        createImportResolutionKey(resolution.importerId, resolution.importPath),
-        {
-          resolvedId: resolution.resolvedId,
-          sourceMetadata: createStaticCssEvalProviderSourceMetadata(resolution)
-        }
-      );
-    }
-  }
-
-  resolve(query: StaticCssEvalQuery): StaticCssEvalResult {
-    if (!query.bindingName) {
-      return { kind: "not-candidate" };
-    }
-
-    const ownerRecord = this.#getModuleRecord(query.importerId);
-
-    if (!ownerRecord) {
-      return { kind: "not-candidate" };
-    }
-
-    const owner = createQueryOwnerLocation(query);
-    const state = createResolutionState(owner);
-    const importBinding = ownerRecord.imports.get(query.bindingName);
-    const cjsBinding = ownerRecord.cjsImports.get(query.bindingName);
-    const requestResult = importBinding
-      ? createResolutionRequest(query, importBinding)
-      : cjsBinding
-        ? createCjsResolutionRequest(query, cjsBinding)
-        : null;
-
-    if (!requestResult) {
-      const cjsDiagnostic = this.#createCjsUnsupportedDiagnostic(
-        ownerRecord,
-        query
-      );
-
-      return cjsDiagnostic
-        ? createImportedStaticCssEvalErrorResult({
-            query,
-            diagnostic: cjsDiagnostic,
-            state
-          })
-        : { kind: "not-candidate" };
-    }
-
-    if (requestResult.kind === "error") {
-      return createImportedStaticCssEvalErrorResult({
-        query,
-        diagnostic: requestResult.diagnostic,
-        state
-      });
-    }
-
-    const importResult = this.#resolveProviderImport(
-      requestResult.request,
-      owner,
-      state
-    );
-
-    if (importResult.kind === "error") {
-      return createImportedStaticCssEvalErrorResult({
-        query,
-        diagnostic: importResult.diagnostic,
-        state
-      });
-    }
-
-    const result = this.#resolveModuleExport(
-      importResult.record,
-      requestResult.request,
-      state
-    );
-
-    if (result.kind === "not-candidate") {
-      return { kind: "not-candidate" };
-    }
-
-    if (result.kind === "error") {
-      return createImportedStaticCssEvalErrorResult({
-        query,
-        diagnostic: result.diagnostic,
-        state,
-        provenance: result.provenance,
-        cacheKey: result.cacheKey
-      });
-    }
-
-    markResolutionDependenciesContributed(state);
-
-    return createImportedStaticCssEvalResolvedResult({
-      query,
-      value: result.value,
-      provenance: result.provenance,
-      cacheKey: result.cacheKey,
-      state
-    });
-  }
-
-  #resolveProviderImport(
-    request: ImportedStaticCssEvalResolutionRequest,
-    owner: StaticCssEvalSourceLocation,
-    state: ImportedStaticCssEvalResolutionState
-  ):
-    | { kind: "resolved"; record: ImportedStaticCssEvalModuleRecord }
-    | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-    const resolvedImport = this.#resolveImport(
-      request.importer,
-      request.specifier
-    );
-
-    if (!resolvedImport) {
-      const sourceMetadata = createStaticCssEvalProviderSourceMetadata({
-        sourceKind: "unresolved",
-        sourceOrigin: "unresolved",
-        unsupportedReason: "unresolved"
-      });
-      addResolutionDependency(state, {
-        file: request.specifier,
-        kind: "unresolved",
-        importer: request.importer,
-        specifier: request.specifier,
-        exportName: request.exportName,
-        memberPath: request.memberPath,
-        inspected: false,
-        contributed: false,
-        ...sourceMetadata
-      });
-
-      return {
-        kind: "error",
-        diagnostic: createStaticCssEvalUnresolvedImportDiagnostic(
-          {
-            owner,
-            dependency: { file: request.specifier },
-            importPath: request.specifier,
-            exportName: request.exportName,
-            memberPath: request.memberPath,
-            importChain: createResolutionImportChain(state, request.specifier)
-          },
-          request.specifier
-        )
-      };
-    }
-
-    const loadedModule = this.#modules.get(resolvedImport.resolvedId);
-    const cachedRecord = this.#records.get(resolvedImport.resolvedId);
-    const resolvedSource = loadedModule ?? cachedRecord;
-    const sourceMetadata = mergeImportedStaticCssEvalSourceMetadata(
-      resolvedImport.sourceMetadata,
-      resolvedSource
-    );
-
-    addResolutionDependency(state, {
-      file: resolvedImport.resolvedId,
-      kind: request.dependencyKind,
-      importer: request.importer,
-      specifier: request.specifier,
-      exportName: request.exportName,
-      memberPath: request.memberPath,
-      inspected: false,
-      contributed: false,
-      ...sourceMetadata
-    });
-
-    const sourcePolicyResult = enforceStaticCssEvalProviderSourcePolicy({
-      owner,
-      dependency: { file: resolvedImport.resolvedId },
-      importPath: request.specifier,
-      exportName: request.exportName,
-      memberPath: request.memberPath,
-      importChain: createResolutionImportChain(
-        state,
-        resolvedImport.resolvedId
-      ),
-      sourceId: resolvedImport.resolvedId,
-      ...sourceMetadata
-    });
-
-    if (!sourcePolicyResult.ok) {
-      markResolutionDependencyInspected(state, resolvedImport.resolvedId);
-      if (sourceMetadata.sourceKind === "unresolved") {
-        return {
-          kind: "error",
-          diagnostic: createStaticCssEvalUnresolvedImportDiagnostic(
-            {
-              owner,
-              dependency: { file: resolvedImport.resolvedId },
-              importPath: request.specifier,
-              exportName: request.exportName,
-              memberPath: request.memberPath,
-              importChain: createResolutionImportChain(
-                state,
-                resolvedImport.resolvedId
-              )
-            },
-            request.specifier
-          )
-        };
-      }
-      return { kind: "error", diagnostic: sourcePolicyResult.diagnostic };
-    }
-
-    if (!resolvedSource) {
-      markResolutionDependencyInspected(state, resolvedImport.resolvedId);
-
-      if (sourceMetadata.sourceKind === "provider-virtual") {
-        return {
-          kind: "error",
-          diagnostic: createStaticCssEvalProviderSourceUnsupportedDiagnostic({
-            owner,
-            dependency: { file: resolvedImport.resolvedId },
-            importPath: request.specifier,
-            exportName: request.exportName,
-            memberPath: request.memberPath,
-            importChain: createResolutionImportChain(
-              state,
-              resolvedImport.resolvedId
-            ),
-            sourceId: resolvedImport.resolvedId,
-            sourceKind: sourceMetadata.sourceKind,
-            sourceOrigin: sourceMetadata.sourceOrigin,
-            reason:
-              sourceMetadata.unsupportedReason ?? "provider-virtual-no-source"
-          })
-        };
-      }
-
-      return {
-        kind: "error",
-        diagnostic: createStaticCssEvalUnresolvedImportDiagnostic(
-          {
-            owner,
-            dependency: { file: resolvedImport.resolvedId },
-            importPath: request.specifier,
-            exportName: request.exportName,
-            memberPath: request.memberPath,
-            importChain: createResolutionImportChain(
-              state,
-              resolvedImport.resolvedId
-            )
-          },
-          request.specifier
-        )
-      };
-    }
-
-    const sourceSizeResult = enforceStaticCssEvalSourceSize({
-      owner,
-      dependency: { file: resolvedImport.resolvedId },
-      importPath: request.specifier,
-      exportName: request.exportName,
-      memberPath: request.memberPath,
-      source: resolvedSource.source
-    });
-
-    if (!sourceSizeResult.ok) {
-      markResolutionDependencyInspected(state, resolvedImport.resolvedId);
-      return { kind: "error", diagnostic: sourceSizeResult.diagnostic };
-    }
-
-    const record =
-      cachedRecord ?? this.#getModuleRecord(resolvedImport.resolvedId);
-
-    if (!record) {
-      markResolutionDependencyInspected(state, resolvedImport.resolvedId);
-      return {
-        kind: "error",
-        diagnostic: createStaticCssEvalUnresolvedImportDiagnostic(
-          {
-            owner,
-            dependency: { file: resolvedImport.resolvedId },
-            importPath: request.specifier,
-            exportName: request.exportName,
-            memberPath: request.memberPath,
-            importChain: createResolutionImportChain(
-              state,
-              resolvedImport.resolvedId
-            )
-          },
-          request.specifier,
-          this.#parseFailures.get(resolvedImport.resolvedId)
-        )
-      };
-    }
-
-    markResolutionDependencyInspected(state, resolvedImport.resolvedId);
-    return { kind: "resolved", record };
-  }
-
-  #resolveModuleExport(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ): ImportedStaticCssEvalResolutionResult {
-    const guardResult = this.#guardResolution(record, request, state);
-
-    if (!guardResult.ok) {
-      addResolutionGuardDependencies(state, guardResult.dependencies, request);
-
-      return { kind: "error", diagnostic: guardResult.diagnostic };
-    }
-
-    const stackEntry = createImportCycleStackEntry(record, request);
-    state.stack.push(stackEntry);
-
-    try {
-      return this.#resolveGuardedModuleExport(record, request, state);
-    } finally {
-      state.stack.pop();
-    }
-  }
-
-  #resolveGuardedModuleExport(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ): ImportedStaticCssEvalResolutionResult {
-    if (request.wholeNamespace === true) {
-      return this.#resolveWholeNamespaceExport(record, request, state);
-    }
-
-    const exportEntry = record.exports.get(request.exportName);
-
-    if (!exportEntry) {
-      const cjsModuleUnsupportedEntry = getCjsModuleUnsupportedEntry(record);
-
-      if (cjsModuleUnsupportedEntry) {
-        return this.#resolveExportMapEntry(
-          record,
-          cjsModuleUnsupportedEntry,
-          request,
-          state
-        );
-      }
-
-      return this.#resolveExportStarEntries(record, request, state);
-    }
-
-    return this.#resolveExportMapEntry(record, exportEntry, request, state);
-  }
-
-  #resolveExportMapEntry(
-    record: ImportedStaticCssEvalModuleRecord,
-    exportEntry: ExportMapEntry,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ): ImportedStaticCssEvalResolutionResult {
-    const cjsReexportSource =
-      exportEntry.kind === "reexport" || exportEntry.kind === "unsupported"
-        ? null
-        : getStaticExportEntryCjsRequireSource(record, exportEntry);
-    const effectiveRequest =
-      exportEntry.kind === "reexport" || cjsReexportSource
-        ? {
-            ...request,
-            dependencyKind: "reexported" as const,
-            provenanceKind: "reexported" as const,
-            reexportName: formatExportName(exportEntry.exportName)
-          }
-        : request;
-    const provenance = createBindingProvenance(record.id, effectiveRequest);
-    state.resolutionChain.push({
-      importer: effectiveRequest.importer,
-      source: record.id,
-      exportName: effectiveRequest.exportName,
-      memberPath: [...effectiveRequest.memberPath],
-      provenance,
-      ...createStaticCssEvalProviderSourceMetadata(record)
-    });
-    updateResolutionDependencyKind(
-      state,
-      record.id,
-      effectiveRequest.dependencyKind
-    );
-
-    if (exportEntry.kind === "unsupported") {
-      return {
-        kind: "error",
-        diagnostic: createUnsupportedExportEntryDiagnostic(
-          exportEntry,
-          record,
-          effectiveRequest,
-          state
-        ),
-        provenance,
-        cacheKey: createImportedStaticCssEvalCacheKey(
-          state.owner.file,
-          record.parsedModule,
-          effectiveRequest.exportName,
-          effectiveRequest.memberPath,
-          record
-        )
-      };
-    }
-
-    if (exportEntry.kind === "reexport") {
-      return this.#resolveReexportEntry(
-        record,
-        exportEntry,
-        effectiveRequest,
-        state,
-        provenance
-      );
-    }
-
-    if (cjsReexportSource) {
-      return this.#resolveCjsReexportEntry(
-        record,
-        exportEntry,
-        cjsReexportSource,
-        effectiveRequest,
-        state,
-        provenance
-      );
-    }
-
-    return resolveStaticExportMapEntry(
-      record,
-      exportEntry,
-      effectiveRequest,
-      state,
-      provenance,
-      (options) => this.#resolveStaticCssLiteralReference(options)
-    );
-  }
-
-  #resolveWholeNamespaceExport(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ): ImportedStaticCssEvalResolutionResult {
-    const namespaceTableResult = this.#collectNamespaceExportNameTable(
-      record,
-      request,
-      state,
-      true
-    );
-    const provenance = createBindingProvenance(record.id, request);
-    const cacheKey = createImportedStaticCssEvalCacheKey(
-      state.owner.file,
-      record.parsedModule,
-      request.exportName,
-      request.memberPath,
-      record
-    );
-
-    if (namespaceTableResult.kind === "error") {
-      return {
-        kind: "error",
-        diagnostic: namespaceTableResult.diagnostic,
-        provenance,
-        cacheKey
-      };
-    }
-
-    const namespaceValue: Record<string, StaticCssLiteral> =
-      Object.create(null);
-
-    for (const [exportName, tableEntry] of namespaceTableResult.table.table) {
-      if (exportName === null) {
-        continue;
-      }
-
-      const exportRequest: ImportedStaticCssEvalResolutionRequest = {
-        ...request,
-        exportName,
-        memberPath: [],
-        wholeNamespace: false,
-        reexportName: formatExportName(exportName)
-      };
-      const exportResult = this.#resolveNamespaceTableEntry(
-        record,
-        tableEntry,
-        exportRequest,
-        namespaceTableResult.table.candidates,
-        state
-      );
-
-      if (exportResult.kind === "not-candidate") {
-        return this.#createMissingExportResult(record, exportRequest, state);
-      }
-
-      if (exportResult.kind === "error") {
-        return {
-          kind: "error",
-          diagnostic: createStaticCssEvalPartialNamespaceFailureDiagnostic({
-            owner: state.owner,
-            dependency: exportResult.diagnostic.dependency ?? {
-              file: record.id
-            },
-            importPath: request.specifier,
-            exportName: exportRequest.exportName,
-            memberPath: exportRequest.memberPath,
-            importChain: exportResult.diagnostic.importChain,
-            failedReason: exportResult.diagnostic.reason
-          }),
-          provenance,
-          cacheKey
-        };
-      }
-
-      namespaceValue[exportName] = exportResult.value;
-    }
-
-    return {
-      kind: "resolved",
-      value: namespaceValue,
-      provenance,
-      cacheKey
-    };
-  }
-
-  #resolveNamespaceTableEntry(
-    record: ImportedStaticCssEvalModuleRecord,
-    tableEntry: StaticCssModuleExportNameTableEntry,
-    request: ImportedStaticCssEvalResolutionRequest,
-    candidates: readonly ImportedStaticCssEvalExportStarCandidate[],
-    state: ImportedStaticCssEvalResolutionState
-  ): ImportedStaticCssEvalResolutionResult {
-    switch (tableEntry.kind) {
-      case "explicit":
-        return this.#resolveExportMapEntry(
-          record,
-          tableEntry.entry,
-          request,
-          state
-        );
-      case "star": {
-        const candidate = candidates.find(
-          (exportStarCandidate) =>
-            exportStarCandidate.entry === tableEntry.entry
-        );
-
-        return candidate
-          ? this.#resolveModuleExport(
-              candidate.record,
-              {
-                ...candidate.request,
-                exportName: request.exportName,
-                memberPath: [],
-                wholeNamespace: false,
-                reexportName: formatExportName(request.exportName)
-              },
-              state
-            )
-          : this.#createMissingExportResult(record, request, state);
-      }
-      case "ambiguous-star":
-        return {
-          kind: "error",
-          diagnostic: createAmbiguousExportStarDiagnostic(
-            record,
-            request,
-            state,
-            candidates
-          ),
-          cacheKey: createImportedStaticCssEvalCacheKey(
-            state.owner.file,
-            record.parsedModule,
-            request.exportName,
-            request.memberPath,
-            record
-          )
-        };
-      default:
-        return assertNever(tableEntry);
-    }
-  }
-
-  #collectNamespaceExportNames(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState,
-    includeDefaultExport: boolean
-  ):
-    | { kind: "resolved"; exportNames: StaticCssEvalExportName[] }
-    | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-    const guardResult = this.#guardResolution(record, request, state);
-
-    if (!guardResult.ok) {
-      addResolutionGuardDependencies(state, guardResult.dependencies, request);
-
-      return { kind: "error", diagnostic: guardResult.diagnostic };
-    }
-
-    const stackEntry = createImportCycleStackEntry(record, request);
-    state.stack.push(stackEntry);
-
-    try {
-      const namespaceTableResult = this.#collectNamespaceExportNameTable(
-        record,
-        request,
-        state,
-        includeDefaultExport
-      );
-
-      if (namespaceTableResult.kind === "error") {
-        return namespaceTableResult;
-      }
-
-      const exportNames: StaticCssEvalExportName[] = [];
-
-      for (const [exportName, tableEntry] of namespaceTableResult.table.table) {
-        if (exportName === null) {
-          continue;
-        }
-
-        if (tableEntry.kind === "ambiguous-star") {
-          return {
-            kind: "error",
-            diagnostic: createAmbiguousExportStarDiagnostic(
-              record,
-              { ...request, exportName },
-              state,
-              namespaceTableResult.table.candidates
-            )
-          };
-        }
-
-        exportNames.push(exportName);
-      }
-
-      return { kind: "resolved", exportNames };
-    } finally {
-      state.stack.pop();
-    }
-  }
-
-  #collectNamespaceExportNameTable(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState,
-    includeDefaultExport: boolean
-  ):
-    | { kind: "resolved"; table: ImportedStaticCssEvalNamespaceExportNameTable }
-    | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-    const starSourcesResult = this.#collectNamespaceExportStarSources(
-      record,
-      request,
-      state
-    );
-
-    if (starSourcesResult.kind === "error") {
-      return { kind: "error", diagnostic: starSourcesResult.diagnostic };
-    }
-
-    return {
-      kind: "resolved",
-      table: {
-        table: createStaticCssModuleExportNameTable(
-          createNamespaceExplicitExportMap(
-            record.exports,
-            includeDefaultExport
-          ),
-          starSourcesResult.starSources
-        ),
-        candidates: starSourcesResult.candidates
-      }
-    };
-  }
-
-  #collectNamespaceExportStarSources(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ):
-    | {
-        kind: "resolved";
-        starSources: StaticCssModuleExportStarSource[];
-        candidates: ImportedStaticCssEvalExportStarCandidate[];
-      }
-    | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-    const starSources: StaticCssModuleExportStarSource[] = [];
-    const candidates: ImportedStaticCssEvalExportStarCandidate[] = [];
-
-    for (const starEntry of record.parsedModule.exportStarReexports) {
-      const starRequest = createExportStarRequest(record, starEntry, request);
-      const importResult = this.#resolveProviderImport(
-        starRequest,
-        state.owner,
-        state
-      );
-
-      if (importResult.kind === "error") {
-        return { kind: "error", diagnostic: importResult.diagnostic };
-      }
-
-      const exportNamesResult = this.#collectNamespaceExportNames(
-        importResult.record,
-        { ...starRequest, wholeNamespace: true },
-        state,
-        false
-      );
-
-      if (exportNamesResult.kind === "error") {
-        return { kind: "error", diagnostic: exportNamesResult.diagnostic };
-      }
-
-      if (exportNamesResult.exportNames.length === 0) {
-        continue;
-      }
-
-      candidates.push({
-        entry: starEntry,
-        record: importResult.record,
-        request: starRequest
-      });
-      starSources.push({
-        entry: starEntry,
-        exportNames: exportNamesResult.exportNames
-      });
-    }
-
-    return { kind: "resolved", starSources, candidates };
-  }
-
-  #resolveExportStarEntries(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ): ImportedStaticCssEvalResolutionResult {
-    if (
-      request.exportName === "default" ||
-      record.parsedModule.exportStarReexports.length === 0
-    ) {
-      return this.#createMissingExportResult(record, request, state);
-    }
-
-    const effectiveRequest = {
-      ...request,
-      dependencyKind: "reexported" as const,
-      provenanceKind: "reexported" as const,
-      reexportName: formatExportName(request.exportName)
-    };
-    const provenance = createBindingProvenance(record.id, effectiveRequest);
-    state.resolutionChain.push({
-      importer: effectiveRequest.importer,
-      source: record.id,
-      exportName: effectiveRequest.exportName,
-      memberPath: [...effectiveRequest.memberPath],
-      provenance,
-      ...createStaticCssEvalProviderSourceMetadata(record)
-    });
-    updateResolutionDependencyKind(
-      state,
-      record.id,
-      effectiveRequest.dependencyKind
-    );
-
-    const candidatesResult = this.#collectExportStarCandidates(
-      record,
-      effectiveRequest,
-      state
-    );
-
-    if (candidatesResult.kind === "error") {
-      return {
-        kind: "error",
-        diagnostic: candidatesResult.diagnostic,
-        provenance,
-        cacheKey: createImportedStaticCssEvalCacheKey(
-          state.owner.file,
-          record.parsedModule,
-          effectiveRequest.exportName,
-          effectiveRequest.memberPath,
-          record
-        )
-      };
-    }
-
-    const exportNameTable = createStaticCssModuleExportNameTable(
-      record.exports,
-      createExportStarSources(
-        effectiveRequest.exportName,
-        candidatesResult.candidates
-      )
-    );
-    const tableEntry = exportNameTable.get(effectiveRequest.exportName);
-
-    if (!tableEntry) {
-      return this.#createMissingExportResult(record, effectiveRequest, state);
-    }
-
-    switch (tableEntry.kind) {
-      case "explicit":
-        return this.#resolveExportMapEntry(
-          record,
-          tableEntry.entry,
-          effectiveRequest,
-          state
-        );
-      case "star": {
-        const candidate = candidatesResult.candidates.find(
-          (exportStarCandidate) =>
-            exportStarCandidate.entry === tableEntry.entry
-        );
-
-        return candidate
-          ? this.#resolveModuleExport(
-              candidate.record,
-              candidate.request,
-              state
-            )
-          : this.#createMissingExportResult(record, effectiveRequest, state);
-      }
-      case "ambiguous-star":
-        return {
-          kind: "error",
-          diagnostic: createAmbiguousExportStarDiagnostic(
-            record,
-            effectiveRequest,
-            state,
-            candidatesResult.candidates
-          ),
-          provenance,
-          cacheKey: createImportedStaticCssEvalCacheKey(
-            state.owner.file,
-            record.parsedModule,
-            effectiveRequest.exportName,
-            effectiveRequest.memberPath,
-            record
-          )
-        };
-      default:
-        return assertNever(tableEntry);
-    }
-  }
-
-  #collectExportStarCandidates(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ):
-    | {
-        kind: "resolved";
-        candidates: ImportedStaticCssEvalExportStarCandidate[];
-      }
-    | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-    const candidates: ImportedStaticCssEvalExportStarCandidate[] = [];
-
-    for (const starEntry of record.parsedModule.exportStarReexports) {
-      const candidateResult = this.#resolveExportStarCandidate(
-        record,
-        starEntry,
-        request,
-        state
-      );
-
-      if (candidateResult.kind === "error") {
-        return { kind: "error", diagnostic: candidateResult.diagnostic };
-      }
-
-      if (candidateResult.kind === "resolved") {
-        candidates.push(candidateResult.candidate);
-      }
-    }
-
-    return { kind: "resolved", candidates };
-  }
-
-  #resolveExportStarCandidate(
-    record: ImportedStaticCssEvalModuleRecord,
-    starEntry: ExportGraphStarReexportEntry,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ):
-    | {
-        kind: "resolved";
-        candidate: ImportedStaticCssEvalExportStarCandidate;
-      }
-    | { kind: "missing" }
-    | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-    const starRequest = createExportStarRequest(record, starEntry, request);
-    const importResult = this.#resolveProviderImport(
-      starRequest,
-      state.owner,
-      state
-    );
-
-    if (importResult.kind === "error") {
-      return { kind: "error", diagnostic: importResult.diagnostic };
-    }
-
-    const presenceResult = this.#recordExportsName(
-      importResult.record,
-      starRequest,
-      state
-    );
-
-    if (presenceResult.kind === "error") {
-      return { kind: "error", diagnostic: presenceResult.diagnostic };
-    }
-
-    return presenceResult.kind === "found"
-      ? {
-          kind: "resolved",
-          candidate: {
-            entry: starEntry,
-            record: importResult.record,
-            request: starRequest
-          }
-        }
-      : { kind: "missing" };
-  }
-
-  #recordExportsName(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ): ImportedStaticCssEvalExportPresenceResult {
-    const guardResult = this.#guardResolution(record, request, state);
-
-    if (!guardResult.ok) {
-      addResolutionGuardDependencies(state, guardResult.dependencies, request);
-
-      return { kind: "error", diagnostic: guardResult.diagnostic };
-    }
-
-    const stackEntry = createImportCycleStackEntry(record, request);
-    state.stack.push(stackEntry);
-
-    try {
-      if (record.exports.has(request.exportName)) {
-        return { kind: "found" };
-      }
-
-      if (
-        request.exportName === "default" ||
-        record.parsedModule.exportStarReexports.length === 0
-      ) {
-        return { kind: "missing" };
-      }
-
-      const candidatesResult = this.#collectExportStarCandidates(
-        record,
-        request,
-        state
-      );
-
-      if (candidatesResult.kind === "error") {
-        return { kind: "error", diagnostic: candidatesResult.diagnostic };
-      }
-
-      const exportNameTable = createStaticCssModuleExportNameTable(
-        record.exports,
-        createExportStarSources(request.exportName, candidatesResult.candidates)
-      );
-      const tableEntry = exportNameTable.get(request.exportName);
-
-      if (!tableEntry) {
-        return { kind: "missing" };
-      }
-
-      switch (tableEntry.kind) {
-        case "explicit":
-        case "star":
-          return { kind: "found" };
-        case "ambiguous-star":
-          return {
-            kind: "error",
-            diagnostic: createAmbiguousExportStarDiagnostic(
-              record,
-              request,
-              state,
-              candidatesResult.candidates
-            )
-          };
-        default:
-          return assertNever(tableEntry);
-      }
-    } finally {
-      state.stack.pop();
-    }
-  }
-
-  #resolveReexportEntry(
-    record: ImportedStaticCssEvalModuleRecord,
-    exportEntry: Extract<ExportMapEntry, { kind: "reexport" }>,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState,
-    provenance: BindingProvenance
-  ): ImportedStaticCssEvalResolutionResult {
-    const reexportRequest: ImportedStaticCssEvalResolutionRequest = {
-      importer: record.id,
-      specifier: exportEntry.source,
-      exportName: exportEntry.importedName,
-      memberPath: [...request.memberPath],
-      dependencyKind: "reexported",
-      provenanceKind: "reexported",
-      reexportName: formatExportName(exportEntry.exportName)
-    };
-    const importResult = this.#resolveProviderImport(
-      reexportRequest,
-      state.owner,
-      state
-    );
-
-    return importResult.kind === "error"
-      ? {
-          kind: "error",
-          diagnostic: importResult.diagnostic,
-          provenance,
-          cacheKey: createImportedStaticCssEvalCacheKey(
-            state.owner.file,
-            record.parsedModule,
-            request.exportName,
-            request.memberPath,
-            record
-          )
-        }
-      : this.#resolveModuleExport(importResult.record, reexportRequest, state);
-  }
-
-  #resolveCjsReexportEntry(
-    record: ImportedStaticCssEvalModuleRecord,
-    exportEntry: Exclude<ExportMapEntry, { kind: "reexport" | "unsupported" }>,
-    source: StaticCssEvalCjsRequireSource,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState,
-    provenance: BindingProvenance
-  ): ImportedStaticCssEvalResolutionResult {
-    const reexportParts = createStaticCssEvalCjsReexportRequestParts({
-      source,
-      queryMemberPath: request.memberPath,
-      reexportName: formatExportName(exportEntry.exportName)
-    });
-    const reexportRequest = createResolutionRequestFromCjsParts(
-      record.id,
-      reexportParts
-    );
-    const importResult = this.#resolveProviderImport(
-      reexportRequest,
-      state.owner,
-      state
-    );
-
-    return importResult.kind === "error"
-      ? {
-          kind: "error",
-          diagnostic: importResult.diagnostic,
-          provenance,
-          cacheKey: createImportedStaticCssEvalCacheKey(
-            state.owner.file,
-            record.parsedModule,
-            request.exportName,
-            request.memberPath,
-            record
-          )
-        }
-      : this.#resolveModuleExport(importResult.record, reexportRequest, state);
-  }
-
-  #createMissingExportResult(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ): ImportedStaticCssEvalResolutionResult {
-    return {
-      kind: "error",
-      diagnostic: createStaticCssEvalUnresolvedExportDiagnostic(
-        {
-          owner: state.owner,
-          dependency: { file: record.id },
-          importPath: request.specifier,
-          exportName: request.exportName,
-          memberPath: request.memberPath,
-          importChain: createResolutionImportChain(state, record.id)
-        },
-        request.exportName
-      ),
-      cacheKey: createImportedStaticCssEvalCacheKey(
-        state.owner.file,
-        record.parsedModule,
-        request.exportName,
-        request.memberPath,
-        record
-      )
-    };
-  }
-
-  #guardResolution(
-    record: ImportedStaticCssEvalModuleRecord,
-    request: ImportedStaticCssEvalResolutionRequest,
-    state: ImportedStaticCssEvalResolutionState
-  ) {
-    const next = {
-      file: record.id,
-      exportName: request.exportName,
-      memberPath: request.memberPath
-    };
-    const cycleResult = guardStaticCssEvalImportCycle({
-      owner: state.owner,
-      dependency: { file: record.id },
-      importPath: request.specifier,
-      exportName: request.exportName,
-      memberPath: request.memberPath,
-      stack: state.stack,
-      next
-    });
-
-    if (!cycleResult.ok) {
-      return cycleResult;
-    }
-
-    return guardStaticCssEvalResolutionDepth({
-      owner: state.owner,
-      dependency: { file: record.id },
-      importPath: request.specifier,
-      exportName: request.exportName,
-      memberPath: request.memberPath,
-      resolutionDepth: state.stack.length + 1
-    });
-  }
-
-  #resolveStaticCssLiteralReference(
-    options: ImportedStaticCssEvalLiteralReferenceOptions
-  ): ImportedStaticCssEvalLiteralResult {
-    const { reference } = options;
-
-    if (reference.kind === "unsupported") {
-      return createUnsupportedImportedReferenceResult(options);
-    }
-
-    const cjsBinding = options.record.cjsImports.get(reference.bindingName);
-
-    if (cjsBinding) {
-      return this.#resolveProviderBackedLiteralReference(
-        options,
-        createCjsResolutionRequest(
-          createReferenceQuery(options.record.id, reference),
-          cjsBinding
-        ).request
-      );
-    }
-
-    const dynamicRequireDiagnostic =
-      createDynamicRequireOperandDiagnostic(options);
-
-    if (dynamicRequireDiagnostic) {
-      return { kind: "error", diagnostic: dynamicRequireDiagnostic };
-    }
-
-    const localResult = resolveSameModuleStaticLiteralReference(options);
-
-    if (localResult.kind !== "not-candidate") {
-      return localResult;
-    }
-
-    const importBinding = options.record.imports.get(reference.bindingName);
-
-    if (!importBinding) {
-      return createUnsupportedImportedLiteralResult(
-        options,
-        options.expression
-      );
-    }
-
-    const requestResult = createResolutionRequest(
-      createReferenceQuery(options.record.id, reference),
-      importBinding
-    );
-
-    if (requestResult.kind === "error") {
-      return { kind: "error", diagnostic: requestResult.diagnostic };
-    }
-
-    return this.#resolveProviderBackedLiteralReference(
-      options,
-      requestResult.request
-    );
-  }
-
-  #resolveProviderBackedLiteralReference(
-    options: ImportedStaticCssEvalLiteralReferenceOptions,
-    request: ImportedStaticCssEvalResolutionRequest
-  ): ImportedStaticCssEvalLiteralResult {
-    const importResult = this.#resolveProviderImport(
-      request,
-      options.resolutionState.owner,
-      options.resolutionState
-    );
-
-    if (importResult.kind === "error") {
-      return { kind: "error", diagnostic: importResult.diagnostic };
-    }
-
-    const exportResult = this.#resolveModuleExport(
-      importResult.record,
-      request,
-      options.resolutionState
-    );
-
-    if (exportResult.kind === "error") {
-      return { kind: "error", diagnostic: exportResult.diagnostic };
-    }
-
-    if (exportResult.kind === "not-candidate") {
-      return createUnsupportedImportedLiteralResult(
-        options,
-        options.expression
-      );
-    }
-
-    return { kind: "resolved", value: exportResult.value };
-  }
-
-  #createCjsUnsupportedDiagnostic(
-    ownerRecord: ImportedStaticCssEvalModuleRecord,
-    query: StaticCssEvalQuery
-  ): StaticCssEvalDiagnostic | null {
-    const bindingName = query.bindingName;
-
-    if (!bindingName) {
-      return null;
-    }
-
-    const binding = ownerRecord.programPath.scope.getBinding(bindingName);
-    const init = binding
-      ? getStaticCssEvalConstBindingInitExpression(binding)
-      : null;
-
-    const unwrappedInit = init
-      ? unwrapTransparentCssRuleExpression(init)
-      : null;
-
-    if (
-      !unwrappedInit ||
-      !containsStaticCssEvalRequireCallExpression(
-        unwrappedInit,
-        ownerRecord.programPath.scope
-      )
-    ) {
-      return null;
-    }
-
-    const context = {
-      owner: createQueryOwnerLocation(query),
-      ...(query.memberPath ? { memberPath: query.memberPath } : {})
-    };
-
-    if (
-      !isStaticCssEvalLiteralRequireCallExpression(
-        unwrappedInit,
-        ownerRecord.programPath.scope
-      )
-    ) {
-      return createStaticCssEvalCjsDynamicRequireUnsupportedDiagnostic(context);
-    }
-
-    return createStaticCssEvalCjsUnsupportedDiagnostic(context);
-  }
-
-  #resolveImport(
-    importerId: string,
-    importPath: string
-  ): ImportedStaticCssEvalResolvedImport | null {
-    const resolvedImport = this.#importResolutions.get(
-      createImportResolutionKey(importerId, importPath)
-    );
-
-    if (resolvedImport) {
-      return resolvedImport;
-    }
-
-    const loadedModule = this.#modules.get(importPath);
-
-    return loadedModule
-      ? {
-          resolvedId: importPath,
-          sourceMetadata:
-            createStaticCssEvalProviderSourceMetadata(loadedModule)
-        }
-      : null;
-  }
-
-  #getModuleRecord(id: string): ImportedStaticCssEvalModuleRecord | null {
-    const cachedRecord = this.#records.get(id);
-
-    if (cachedRecord) {
-      return cachedRecord;
-    }
-
-    const loadedModule = this.#modules.get(id);
-
-    if (!loadedModule) {
-      return null;
-    }
-
-    try {
-      const record = createImportedStaticCssEvalModuleRecordWithCache(
-        loadedModule,
-        this.#moduleCache
-      );
-      this.#records.set(id, record);
-      return record;
-    } catch (error) {
-      this.#parseFailures.set(
-        id,
-        error instanceof Error ? error.message : String(error)
-      );
-      return null;
-    }
-  }
-}
-
-function collectModuleImportBindings(
-  statement: t.Statement,
-  imports: Map<string, ImportedStaticCssEvalImportBinding>
-): void {
-  if (!t.isImportDeclaration(statement) || statement.importKind === "type") {
-    return;
-  }
-
-  const importPath = statement.source.value;
-
-  for (const specifier of statement.specifiers) {
-    if (t.isImportDefaultSpecifier(specifier)) {
-      imports.set(specifier.local.name, {
-        kind: "default",
-        localName: specifier.local.name,
-        importPath,
-        importedName: "default"
-      });
-      continue;
-    }
-
-    if (t.isImportSpecifier(specifier) && specifier.importKind !== "type") {
-      const importedName = getModuleStringName(specifier.imported);
-
-      if (importedName) {
-        imports.set(specifier.local.name, {
-          kind: "named",
-          localName: specifier.local.name,
-          importPath,
-          importedName
-        });
-      }
-      continue;
-    }
-
-    if (t.isImportNamespaceSpecifier(specifier)) {
-      imports.set(specifier.local.name, {
-        kind: "namespace",
-        localName: specifier.local.name,
-        importPath
-      });
-    }
-  }
-}
-
-function createInlineStaticCssSourceHash(source: string): string {
-  return `inline:${hash(source)}`;
-}
-
-function createStaticCssModuleSource(
-  loadedModule: ImportedStaticCssEvalLoadedModule
-): StaticCssModuleSource {
-  return {
-    resolvedFile: loadedModule.id,
-    source: loadedModule.source,
-    sourceHash:
-      loadedModule.sourceHash ??
-      createInlineStaticCssSourceHash(loadedModule.source),
-    ...(loadedModule.version !== undefined
-      ? { sourceVersion: loadedModule.version }
-      : {})
-  };
-}
-
-function mergeImportedStaticCssEvalSourceMetadata(
-  resolutionMetadata: StaticCssEvalProviderSourceMetadata,
-  loadedModule: StaticCssEvalProviderSourcePolicyDescriptor | undefined
-): StaticCssEvalProviderSourceMetadata {
-  if (!loadedModule) {
-    return resolutionMetadata;
-  }
-
-  return createStaticCssEvalProviderSourceMetadata({
-    sourceKind: loadedModule.sourceKind ?? resolutionMetadata.sourceKind,
-    sourceOrigin:
-      loadedModule.sourceOrigin ??
-      (loadedModule.sourceKind ? undefined : resolutionMetadata.sourceOrigin),
-    canonicalModuleId:
-      loadedModule.canonicalModuleId ?? resolutionMetadata.canonicalModuleId,
-    normalizedPathKey:
-      loadedModule.normalizedPathKey ?? resolutionMetadata.normalizedPathKey,
-    watchFiles: loadedModule.watchFiles ?? resolutionMetadata.watchFiles,
-    unsupportedReason:
-      loadedModule.unsupportedReason ?? resolutionMetadata.unsupportedReason
-  });
-}
-
-function assertNever(value: never): never {
-  throw new TypeError(`Unexpected imported static css eval value: ${value}`);
-}
-
-function createResolutionState(
-  owner: StaticCssEvalSourceLocation
-): ImportedStaticCssEvalResolutionState {
-  return {
-    owner,
-    dependencies: new Map<string, ResolutionDependency>(),
-    resolutionChain: [],
-    stack: []
-  };
-}
-
-function createResolutionRequest(
-  query: StaticCssEvalQuery,
-  importBinding: ImportedStaticCssEvalImportBinding
-):
-  | { kind: "resolved"; request: ImportedStaticCssEvalResolutionRequest }
-  | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-  const queryMemberPath = [...(query.memberPath ?? [])];
-
-  if (importBinding.kind === "namespace") {
-    const [exportName, ...memberPath] = queryMemberPath;
-
-    if (exportName === undefined) {
-      return {
-        kind: "resolved",
-        request: {
-          importer: query.importerId,
-          specifier: importBinding.importPath,
-          exportName: null,
-          memberPath: [],
-          dependencyKind: "namespace-member",
-          provenanceKind: "namespace-member",
-          namespaceBinding: importBinding.localName,
-          wholeNamespace: true
-        }
-      };
-    }
-
-    return {
-      kind: "resolved",
-      request: {
-        importer: query.importerId,
-        specifier: importBinding.importPath,
-        exportName,
-        memberPath,
-        dependencyKind: "namespace-member",
-        provenanceKind: "namespace-member",
-        namespaceBinding: importBinding.localName
-      }
-    };
-  }
-
-  return {
-    kind: "resolved",
-    request: {
-      importer: query.importerId,
-      specifier: importBinding.importPath,
-      exportName: importBinding.importedName,
-      memberPath: queryMemberPath,
-      dependencyKind: "imported",
-      provenanceKind: "imported"
-    }
-  };
-}
-
-function createCjsResolutionRequest(
-  query: StaticCssEvalQuery,
-  binding: ImportedStaticCssEvalCjsBinding
-): { kind: "resolved"; request: ImportedStaticCssEvalResolutionRequest } {
-  const requestParts = createStaticCssEvalCjsImportRequestParts({
-    binding,
-    queryMemberPath: [...(query.memberPath ?? [])]
-  });
-
-  return {
-    kind: "resolved",
-    request: createResolutionRequestFromCjsParts(query.importerId, requestParts)
-  };
-}
-
-function createResolutionRequestFromCjsParts(
-  importer: string,
-  parts: StaticCssEvalCjsResolutionRequestParts
-): ImportedStaticCssEvalResolutionRequest {
-  return {
-    importer,
-    specifier: parts.specifier,
-    exportName: parts.exportName,
-    memberPath: parts.memberPath,
-    dependencyKind: parts.dependencyKind,
-    provenanceKind: parts.provenanceKind,
-    ...(parts.reexportName !== undefined
-      ? { reexportName: parts.reexportName }
-      : {})
-  };
-}
-
-function getStaticExportEntryCjsRequireSource(
-  record: ImportedStaticCssEvalModuleRecord,
-  exportEntry: Exclude<ExportMapEntry, { kind: "reexport" | "unsupported" }>
-): StaticCssEvalCjsRequireSource | null {
-  return exportEntry.kind === "expression"
-    ? getStaticExpressionCjsRequireSource(record, exportEntry.expression)
-    : getLocalBindingCjsRequireSource(record, exportEntry.localName, []);
-}
-
-function getStaticExpressionCjsRequireSource(
-  record: ImportedStaticCssEvalModuleRecord,
-  expression: t.Expression
-): StaticCssEvalCjsRequireSource | null {
-  const unwrappedExpression = unwrapTransparentCssRuleExpression(expression);
-  const directSource = getStaticCssEvalCjsRequireSource(
-    unwrappedExpression,
-    record.programPath.scope
-  );
-
-  if (directSource) {
-    return directSource;
-  }
-
-  const reference = getStaticCssEvalMemberReference(unwrappedExpression);
-
-  return reference?.kind === "supported"
-    ? getLocalBindingCjsRequireSource(
-        record,
-        reference.bindingName,
-        reference.memberPath
-      )
-    : null;
-}
-
-function getLocalBindingCjsRequireSource(
-  record: ImportedStaticCssEvalModuleRecord,
-  localName: string,
-  memberPath: readonly string[]
-): StaticCssEvalCjsRequireSource | null {
-  const binding = record.programPath.scope.getBinding(localName);
-  const expression = binding
-    ? getStaticCssEvalConstBindingInitExpression(binding)
-    : null;
-  const source = getStaticCssEvalCjsRequireSource(
-    expression ? unwrapTransparentCssRuleExpression(expression) : null,
-    record.programPath.scope
-  );
-
-  return source
-    ? appendStaticCssEvalCjsRequireSource(source, memberPath)
-    : null;
-}
-
-function appendStaticCssEvalCjsRequireSource(
-  source: StaticCssEvalCjsRequireSource,
-  memberPath: readonly string[]
-): StaticCssEvalCjsRequireSource {
-  return memberPath.length === 0
-    ? source
-    : {
-        importPath: source.importPath,
-        propertyPath: [...source.propertyPath, ...memberPath]
-      };
-}
-
-function resolveStaticExportMapEntry(
-  record: ImportedStaticCssEvalModuleRecord,
-  exportEntry: Exclude<ExportMapEntry, { kind: "reexport" | "unsupported" }>,
-  request: ImportedStaticCssEvalResolutionRequest,
-  state: ImportedStaticCssEvalResolutionState,
-  provenance: BindingProvenance,
-  resolveReference: ImportedStaticCssEvalLiteralReferenceResolver
-): ImportedStaticCssEvalResolutionResult {
-  const expressionResult = getStaticExportEntryExpression(record, exportEntry);
-  const cacheKey = createImportedStaticCssEvalCacheKey(
-    state.owner.file,
-    record.parsedModule,
-    request.exportName,
-    request.memberPath,
-    record
-  );
-
-  if (expressionResult.kind === "error") {
-    return {
-      kind: "error",
-      diagnostic: expressionResult.diagnostic,
-      provenance,
-      cacheKey
-    };
-  }
-
-  const localName = expressionResult.localName;
-
-  if (localName && hasImportedStaticCssBindingMutation(record, localName)) {
-    return {
-      kind: "error",
-      diagnostic: createStaticCssEvalMutatedBindingDiagnostic(
-        createImportedDiagnosticContext(record, request, state),
-        localName
-      ),
-      provenance,
-      cacheKey
-    };
-  }
-
-  const selectedExpression = selectStaticCssEvalExpressionForMemberPath(
-    expressionResult.expression,
-    request.memberPath
-  );
-
-  const context = createImportedLiteralContext(record, request, state);
-  const literalResult = evaluateStaticCssLiteralExpression({
-    expression: selectedExpression.expression,
-    context,
-    record,
-    request,
-    resolutionState: state,
-    literalState: { count: 0 },
-    localStack: [],
-    depth: 1,
-    resolveReference
-  });
-
-  if (literalResult.kind === "error") {
-    return {
-      kind: "error",
-      diagnostic: literalResult.diagnostic,
-      provenance,
-      cacheKey
-    };
-  }
-
-  const value = selectStaticCssLiteralMemberPath(
-    literalResult.value,
-    selectedExpression.deferredMemberPath
-  );
-
-  if (value.kind === "not-candidate") {
-    return {
-      kind: "error",
-      diagnostic: createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-        createImportedLiteralDiagnosticContext(context),
-        selectedExpression.expression.type
-      ),
-      provenance,
-      cacheKey
-    };
-  }
-
-  return {
-    kind: "resolved",
-    value: value.value,
-    provenance,
-    cacheKey
-  };
-}
-
-function getStaticExportEntryExpression(
-  record: ImportedStaticCssEvalModuleRecord,
-  exportEntry:
-    | ExportMapLocalEntry
-    | Extract<ExportMapEntry, { kind: "expression" }>
-):
-  | { kind: "resolved"; expression: t.Expression; localName: string | null }
-  | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-  if (exportEntry.kind === "expression") {
-    const expression = unwrapTransparentCssRuleExpression(
-      exportEntry.expression
-    );
-    const reference = getStaticCssEvalMemberReference(expression);
-
-    if (reference?.kind === "supported") {
-      return getLocalBindingExpressionResult(
-        record,
-        reference.bindingName,
-        exportEntry.exportName,
-        reference.memberPath
-      );
-    }
-
-    if (reference?.kind === "unsupported") {
-      return {
-        kind: "error",
-        diagnostic: createStaticCssEvalComputedMemberUnsupportedDiagnostic({
-          owner: { file: record.id },
-          dependency: { file: record.id },
-          exportName: exportEntry.exportName,
-          memberPath: reference.memberPath
-        })
-      };
-    }
-
-    return { kind: "resolved", expression, localName: null };
-  }
-
-  return getLocalBindingExpressionResult(
-    record,
-    exportEntry.localName,
-    exportEntry.exportName
-  );
-}
-
-function getLocalBindingExpressionResult(
-  record: ImportedStaticCssEvalModuleRecord,
-  localName: string,
-  exportName: StaticCssEvalExportName,
-  memberPath: readonly string[] = []
-):
-  | { kind: "resolved"; expression: t.Expression; localName: string }
-  | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-  const binding = record.programPath.scope.getBinding(localName);
-  const expression = binding
-    ? getStaticCssEvalConstBindingInitExpression(binding)
-    : null;
-
-  if (expression) {
-    const unwrappedExpression = unwrapTransparentCssRuleExpression(expression);
-    const memberExpression = resolveStaticObjectMemberPath(
-      unwrappedExpression,
-      memberPath
-    );
-
-    if (!memberExpression) {
-      return {
-        kind: "error",
-        diagnostic: createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-          {
-            owner: { file: record.id },
-            dependency: { file: record.id },
-            exportName,
-            memberPath
-          },
-          unwrappedExpression.type
-        )
-      };
-    }
-
-    return {
-      kind: "resolved",
-      expression: memberExpression,
-      localName
-    };
-  }
-
-  return {
-    kind: "error",
-    diagnostic: createStaticCssEvalMutableBindingDiagnostic(
-      {
-        owner: { file: record.id },
-        dependency: { file: record.id },
-        exportName
-      },
-      localName
-    )
-  };
-}
-
-function createImportedLiteralContext(
-  record: ImportedStaticCssEvalModuleRecord,
-  request: ImportedStaticCssEvalResolutionRequest,
-  state: ImportedStaticCssEvalResolutionState
-): ImportedStaticCssEvalContext {
-  return {
-    owner: state.owner,
-    dependency: { file: record.id },
-    importPath: request.specifier,
-    exportName: request.exportName,
-    memberPath: [...request.memberPath]
-  };
-}
-
-function createImportedDiagnosticContext(
-  record: ImportedStaticCssEvalModuleRecord,
-  request: ImportedStaticCssEvalResolutionRequest,
-  state: ImportedStaticCssEvalResolutionState
-) {
-  return {
-    owner: state.owner,
-    dependency: { file: record.id },
-    importPath: request.specifier,
-    exportName: request.exportName,
-    memberPath: request.memberPath,
-    importChain: createResolutionImportChain(state, record.id)
-  };
-}
-
-function createUnsupportedExportEntryDiagnostic(
-  exportEntry: Extract<ExportMapEntry, { kind: "unsupported" }>,
-  record: ImportedStaticCssEvalModuleRecord,
-  request: ImportedStaticCssEvalResolutionRequest,
-  state: ImportedStaticCssEvalResolutionState
-): StaticCssEvalDiagnostic {
-  const context = createImportedDiagnosticContext(record, request, state);
-
-  if (exportEntry.unsupportedKind === "export-star") {
-    return createStaticCssEvalExportStarUnsupportedDiagnostic(
-      context,
-      exportEntry.source ?? request.specifier
-    );
-  }
-
-  if (exportEntry.unsupportedKind === "export-namespace") {
-    return createStaticCssEvalNamespaceReexportUnsupportedDiagnostic(
-      context,
-      exportEntry.source ?? request.specifier
-    );
-  }
-
-  if (exportEntry.unsupportedKind === "cjs-export") {
-    return createStaticCssEvalCjsExportUnsupportedDiagnostic(
-      context,
-      exportEntry.cjsExportMutation ?? exportEntry.declaration.type
-    );
-  }
-
-  if (exportEntry.unsupportedKind === "cjs-helper") {
-    return createStaticCssEvalCjsHelperUnsupportedDiagnostic(
-      context,
-      exportEntry.cjsHelperName ?? exportEntry.declaration.type
-    );
-  }
-
-  if (exportEntry.unsupportedKind === "cjs-bundle-runtime") {
-    return createStaticCssEvalCjsBundleRuntimeUnsupportedDiagnostic(
-      context,
-      exportEntry.cjsBundleRuntimeName ?? exportEntry.declaration.type
-    );
-  }
-
-  return createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-    context,
-    exportEntry.declaration.type
-  );
-}
-
-function getCjsModuleUnsupportedEntry(
-  record: ImportedStaticCssEvalModuleRecord
-): Extract<ExportMapEntry, { kind: "unsupported" }> | null {
-  const entry = record.exports.get(null);
-
-  return entry?.kind === "unsupported" &&
-    (entry.unsupportedKind === "cjs-bundle-runtime" ||
-      entry.unsupportedKind === "cjs-helper" ||
-      entry.unsupportedKind === "cjs-export")
-    ? entry
-    : null;
-}
-
-function createBindingProvenance(
-  file: string,
-  request: ImportedStaticCssEvalResolutionRequest
-): BindingProvenance {
-  if (request.provenanceKind === "namespace-member") {
-    return {
-      kind: "namespace-member",
-      file,
-      importer: request.importer,
-      specifier: request.specifier,
-      exportName: request.exportName,
-      memberPath: [...request.memberPath],
-      namespaceBinding: request.namespaceBinding ?? "<namespace>"
-    };
-  }
-
-  if (request.provenanceKind === "reexported") {
-    return {
-      kind: "reexported",
-      file,
-      importer: request.importer,
-      specifier: request.specifier,
-      exportName: request.exportName,
-      memberPath: [...request.memberPath],
-      reexportName: request.reexportName ?? formatExportName(request.exportName)
-    };
-  }
-
-  return {
-    kind: "imported",
-    file,
-    importer: request.importer,
-    specifier: request.specifier,
-    exportName: request.exportName,
-    memberPath: [...request.memberPath]
-  };
-}
-
-function createImportedStaticCssEvalCacheKey(
-  importerFile: string,
-  parsedModule: ParsedStaticCssModule,
-  exportName: StaticCssEvalExportName,
-  memberPath: readonly string[],
-  sourceMetadata?: StaticCssEvalProviderSourcePolicyDescriptor
-): StaticCssEvalCacheKey {
-  return {
-    importerFile,
-    resolvedFile: parsedModule.resolvedFile,
-    exportName,
-    memberPath: [...memberPath],
-    sourceHash: parsedModule.sourceHash,
-    ...(parsedModule.sourceVersion !== undefined
-      ? { sourceVersion: parsedModule.sourceVersion }
-      : {}),
-    pluginOptionsVersion: "static-css-eval-provider:v1",
-    resolverOptionsVersion: "static-css-eval-provider:v1",
-    parserVersion: parsedModule.cacheKey.parserVersion,
-    staticEvalSupportVersion: STATIC_CSS_MODULE_CACHE_SUPPORT_VERSION,
-    resolvedId: parsedModule.resolvedFile,
-    ...(sourceMetadata
-      ? createStaticCssEvalProviderSourceMetadata(sourceMetadata)
-      : {}),
-    parserOptions: parsedModule.cacheKey.parserOptions
-  };
-}
-
-function createImportedStaticCssEvalResolvedResult(options: {
-  query: StaticCssEvalQuery;
-  value: StaticCssLiteral;
-  provenance: BindingProvenance;
-  cacheKey: StaticCssEvalCacheKey;
-  state: ImportedStaticCssEvalResolutionState;
-}): StaticCssEvalResult {
-  return {
-    kind: "resolved",
-    status: "resolved",
-    expression: formatStaticCssEvalQueryExpression(options.query),
-    value: options.value,
-    provenance: options.provenance,
-    dependencies: getResolutionDependencies(options.state),
-    resolutionChain: [...options.state.resolutionChain],
-    diagnostics: [],
-    cacheKey: options.cacheKey
-  };
-}
-
-function createImportedStaticCssEvalErrorResult(options: {
-  query: StaticCssEvalQuery;
-  diagnostic: StaticCssEvalDiagnostic;
-  state: ImportedStaticCssEvalResolutionState;
-  provenance?: BindingProvenance;
-  cacheKey?: StaticCssEvalCacheKey;
-}): StaticCssEvalResult {
-  return {
-    kind: "error",
-    status: "error",
-    expression: formatStaticCssEvalQueryExpression(options.query),
-    ...(options.provenance ? { provenance: options.provenance } : {}),
-    dependencies: getResolutionDependencies(options.state),
-    resolutionChain: [...options.state.resolutionChain],
-    diagnostic: options.diagnostic,
-    diagnostics: [options.diagnostic],
-    ...(options.cacheKey ? { cacheKey: options.cacheKey } : {})
-  };
-}
-
-function addResolutionDependency(
-  state: ImportedStaticCssEvalResolutionState,
-  dependency: ResolutionDependency
-): void {
-  const key = createResolutionDependencyKey(dependency);
-  const existing = state.dependencies.get(key);
-
-  state.dependencies.set(key, {
-    ...dependency,
-    ...(existing
-      ? {
-          kind: mergeResolutionDependencyKind(existing.kind, dependency.kind),
-          inspected: existing.inspected || dependency.inspected,
-          contributed: existing.contributed || dependency.contributed
-        }
-      : {})
-  });
-}
-
-function updateResolutionDependencyKind(
-  state: ImportedStaticCssEvalResolutionState,
-  file: string,
-  kind: ResolutionDependencyKind
-): void {
-  for (const [key, dependency] of state.dependencies) {
-    if (dependency.file === file) {
-      state.dependencies.set(key, {
-        ...dependency,
-        kind: mergeResolutionDependencyKind(dependency.kind, kind)
-      });
-    }
-  }
-}
-
-function markResolutionDependencyInspected(
-  state: ImportedStaticCssEvalResolutionState,
-  file: string
-): void {
-  for (const [key, dependency] of state.dependencies) {
-    if (dependency.file === file) {
-      state.dependencies.set(key, { ...dependency, inspected: true });
-    }
-  }
-}
-
-function markResolutionDependenciesContributed(
-  state: ImportedStaticCssEvalResolutionState
-): void {
-  for (const [key, dependency] of state.dependencies) {
-    state.dependencies.set(key, {
-      ...dependency,
-      contributed: dependency.inspected && dependency.kind !== "unresolved"
-    });
-  }
-}
-
-function getResolutionDependencies(
-  state: ImportedStaticCssEvalResolutionState
-): ResolutionDependency[] {
-  return [...state.dependencies.values()].map((dependency) => ({
-    ...dependency,
-    memberPath: [...dependency.memberPath],
-    ...(dependency.watchFiles ? { watchFiles: [...dependency.watchFiles] } : {})
-  }));
-}
-
-function createResolutionDependencyKey(
-  dependency: ResolutionDependency
-): string {
-  return JSON.stringify([
-    dependency.file,
-    dependency.importer,
-    dependency.specifier,
-    dependency.exportName,
-    dependency.memberPath
-  ]);
-}
-
-function mergeResolutionDependencyKind(
-  previous: ResolutionDependencyKind,
-  next: ResolutionDependencyKind
-): ResolutionDependencyKind {
-  if (previous === "unresolved" || next === "unresolved") {
-    return next === "unresolved" ? previous : next;
-  }
-
-  if (previous === "reexported" || next === "reexported") {
-    return "reexported";
-  }
-
-  if (previous === "namespace-member" || next === "namespace-member") {
-    return "namespace-member";
-  }
-
-  return next;
-}
-
-function createImportCycleStackEntry(
-  record: ImportedStaticCssEvalModuleRecord,
-  request: ImportedStaticCssEvalResolutionRequest
-): StaticCssEvalImportCycleKey {
-  return {
-    file: record.id,
-    exportName: request.exportName,
-    memberPath: [...request.memberPath]
-  };
-}
-
-function createExportStarRequest(
-  record: ImportedStaticCssEvalModuleRecord,
-  starEntry: ExportGraphStarReexportEntry,
-  request: ImportedStaticCssEvalResolutionRequest
-): ImportedStaticCssEvalResolutionRequest {
-  return {
-    importer: record.id,
-    specifier: starEntry.source,
-    exportName: request.exportName,
-    memberPath: [...request.memberPath],
-    dependencyKind: "reexported",
-    provenanceKind: "reexported",
-    reexportName: formatExportName(request.exportName)
-  };
-}
-
-function createExportStarSources(
-  exportName: StaticCssEvalExportName,
-  candidates: readonly ImportedStaticCssEvalExportStarCandidate[]
-): StaticCssModuleExportStarSource[] {
-  return candidates.map((candidate) => ({
-    entry: candidate.entry,
-    exportNames: [exportName]
-  }));
-}
-
-function createNamespaceExplicitExportMap(
-  exports: ReadonlyMap<StaticCssEvalExportName, ExportMapEntry>,
-  includeDefaultExport: boolean
-): ReadonlyMap<StaticCssEvalExportName, ExportMapEntry> {
-  const namespaceExports = new Map<StaticCssEvalExportName, ExportMapEntry>();
-
-  for (const [exportName, exportEntry] of exports) {
-    if (exportName === null) {
-      continue;
-    }
-
-    if (!includeDefaultExport && exportName === "default") {
-      continue;
-    }
-
-    namespaceExports.set(exportName, exportEntry);
-  }
-
-  return namespaceExports;
-}
-
-function addResolutionGuardDependencies(
-  state: ImportedStaticCssEvalResolutionState,
-  dependencies: readonly string[],
-  request: ImportedStaticCssEvalResolutionRequest
-): void {
-  for (const file of dependencies) {
-    addResolutionDependency(state, {
-      file,
-      kind: request.dependencyKind,
-      importer: request.importer,
-      specifier: request.specifier,
-      exportName: request.exportName,
-      memberPath: [...request.memberPath],
-      inspected: true,
-      contributed: false
-    });
-  }
-}
-
-function createAmbiguousExportStarDiagnostic(
-  record: ImportedStaticCssEvalModuleRecord,
-  request: ImportedStaticCssEvalResolutionRequest,
-  state: ImportedStaticCssEvalResolutionState,
-  candidates: readonly ImportedStaticCssEvalExportStarCandidate[]
-): StaticCssEvalDiagnostic {
-  return createStaticCssEvalAmbiguousExportStarDiagnostic(
-    {
-      owner: state.owner,
-      dependency: { file: record.id },
-      importPath: request.specifier,
-      exportName: request.exportName,
-      memberPath: request.memberPath,
-      importChain: createAmbiguousExportStarImportChain(
-        state,
-        request,
-        candidates
-      )
-    },
-    request.exportName
-  );
-}
-
-function createAmbiguousExportStarImportChain(
-  state: ImportedStaticCssEvalResolutionState,
-  request: ImportedStaticCssEvalResolutionRequest,
-  candidates: readonly ImportedStaticCssEvalExportStarCandidate[]
-): string[] {
-  return [
-    state.owner.file,
-    ...state.stack.map(formatStaticCssEvalResolutionFrame),
-    ...candidates.map(
-      (candidate) =>
-        `${candidate.record.id}#${formatExportName(request.exportName)}`
-    )
-  ];
-}
-
-function createResolutionImportChain(
-  state: ImportedStaticCssEvalResolutionState,
-  nextFile: string
-): string[] {
-  return [
-    state.owner.file,
-    ...state.stack.map(formatStaticCssEvalResolutionFrame),
-    nextFile
-  ];
-}
-
-function formatStaticCssEvalResolutionFrame(
-  frame: StaticCssEvalImportCycleKey
-): string {
-  const memberPath = frame.memberPath?.length
-    ? `.${frame.memberPath.join(".")}`
-    : "";
-
-  return `${frame.file}#${formatExportName(frame.exportName)}${memberPath}`;
-}
-
-function formatStaticCssEvalQueryExpression(query: StaticCssEvalQuery): string {
-  const bindingName = query.bindingName ?? "<unknown>";
-  const memberPath = query.memberPath?.length
-    ? `.${query.memberPath.join(".")}`
-    : "";
-
-  return `${bindingName}${memberPath}`;
-}
-
-function hasImportedStaticCssBindingMutation(
-  record: ImportedStaticCssEvalModuleRecord,
-  localName: string
-): boolean {
-  const binding = record.programPath.scope.getBinding(localName);
-
-  return binding
-    ? hasStaticCssEvalBindingMutation(record.programPath, binding)
-    : false;
-}
-
-function resolveStaticObjectMemberPath(
-  expression: t.Expression,
-  memberPath: readonly string[]
-): t.Expression | null {
-  let currentExpression: t.Expression | null =
-    unwrapTransparentCssRuleExpression(expression);
-
-  for (const memberName of memberPath) {
-    if (!currentExpression) {
-      return null;
-    }
-
-    const unwrappedExpression =
-      unwrapTransparentCssRuleExpression(currentExpression);
-
-    if (!t.isObjectExpression(unwrappedExpression)) {
-      return null;
-    }
-
-    currentExpression = getStaticObjectMemberValue(
-      unwrappedExpression,
-      memberName
-    );
-  }
-
-  return currentExpression
-    ? unwrapTransparentCssRuleExpression(currentExpression)
-    : null;
-}
-
-function selectStaticCssEvalExpressionForMemberPath(
-  expression: t.Expression,
-  memberPath: readonly string[]
-): {
-  readonly expression: t.Expression;
-  readonly deferredMemberPath: string[];
-} {
-  const selectedExpression = resolveStaticObjectMemberPath(
-    expression,
-    memberPath
-  );
-
-  if (selectedExpression) {
-    return {
-      expression: selectedExpression,
-      deferredMemberPath: []
-    };
-  }
-
-  return {
-    expression: unwrapTransparentCssRuleExpression(expression),
-    deferredMemberPath: [...memberPath]
-  };
-}
-
-function selectStaticCssLiteralMemberPath(
-  value: StaticCssLiteral,
-  memberPath: readonly string[]
-): { kind: "resolved"; value: StaticCssLiteral } | { kind: "not-candidate" } {
-  let currentValue = value;
-
-  for (const memberName of memberPath) {
-    if (!isStaticCssLiteralObject(currentValue)) {
-      return { kind: "not-candidate" };
-    }
-
-    if (!Object.prototype.hasOwnProperty.call(currentValue, memberName)) {
-      return { kind: "not-candidate" };
-    }
-
-    const nextValue = currentValue[memberName];
-
-    if (nextValue === undefined) {
-      return { kind: "not-candidate" };
-    }
-
-    currentValue = nextValue;
-  }
-
-  return { kind: "resolved", value: currentValue };
-}
-
-function resolveSameModuleStaticLiteralReference(
-  options: ImportedStaticCssEvalLiteralReferenceOptions
-): ImportedStaticCssEvalLiteralReferenceResult {
-  if (options.reference.kind === "unsupported") {
-    return createUnsupportedImportedReferenceResult(options);
-  }
-
-  const binding = options.record.programPath.scope.getBinding(
-    options.reference.bindingName
-  );
-
-  if (!binding) {
-    return { kind: "not-candidate" };
-  }
-
-  const declarationKind =
-    getImportedStaticCssEvalBindingDeclarationKind(binding);
-  const init = getImportedStaticCssEvalVariableBindingInitExpression(binding);
-
-  if (!declarationKind || !init) {
-    return { kind: "not-candidate" };
-  }
-
-  if (declarationKind === "let" || declarationKind === "var") {
-    return {
-      kind: "error",
-      diagnostic: createStaticCssEvalMutableBindingDiagnostic(
-        createImportedLiteralDiagnosticContext(options.context),
-        options.reference.bindingName
-      )
-    };
-  }
-
-  if (declarationKind !== "const") {
-    return { kind: "not-candidate" };
-  }
-
-  const currentFrame: ImportedStaticCssEvalLocalReferenceFrame = {
-    recordId: options.record.id,
-    bindingName: options.reference.bindingName,
-    memberPath: [...options.reference.memberPath]
-  };
-  const cycleStartIndex = findImportedLocalAliasCycleStartIndex(
-    options.localStack,
-    currentFrame
-  );
-
-  if (cycleStartIndex !== -1) {
-    return {
-      kind: "error",
-      diagnostic: createImportedLocalAliasCycleDiagnostic(options.context, [
-        ...options.localStack.slice(cycleStartIndex),
-        currentFrame
-      ])
-    };
-  }
-
-  if (hasStaticCssEvalBindingMutation(options.record.programPath, binding)) {
-    return {
-      kind: "error",
-      diagnostic: createStaticCssEvalMutatedBindingDiagnostic(
-        createImportedLiteralDiagnosticContext(options.context),
-        options.reference.bindingName
-      )
-    };
-  }
-
-  const selectedExpression = selectStaticCssEvalExpressionForMemberPath(
-    init,
-    options.reference.memberPath
-  );
-  const literalResult = evaluateStaticCssLiteralExpression({
-    ...options,
-    expression: selectedExpression.expression,
-    localStack: [...options.localStack, currentFrame],
-    resolveReference: options.resolveReference
-  });
-
-  if (literalResult.kind === "error") {
-    return literalResult;
-  }
-
-  const selectedValue = selectStaticCssLiteralMemberPath(
-    literalResult.value,
-    selectedExpression.deferredMemberPath
-  );
-
-  return selectedValue.kind === "resolved"
-    ? selectedValue
-    : {
-        kind: "error",
-        diagnostic: createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-          createImportedLiteralDiagnosticContext(options.context),
-          selectedExpression.expression.type
-        )
-      };
-}
-
-function findImportedLocalAliasCycleStartIndex(
-  stack: readonly ImportedStaticCssEvalLocalReferenceFrame[],
-  currentFrame: ImportedStaticCssEvalLocalReferenceFrame
-): number {
-  const currentKey = createImportedLocalAliasCycleKey(currentFrame);
-
-  return stack.findIndex(
-    (frame) => createImportedLocalAliasCycleKey(frame) === currentKey
-  );
-}
-
-function createImportedLocalAliasCycleKey(
-  frame: ImportedStaticCssEvalLocalReferenceFrame
-): string {
-  return JSON.stringify([frame.recordId, frame.bindingName, frame.memberPath]);
-}
-
-function createImportedLocalAliasCycleDiagnostic(
-  context: ImportedStaticCssEvalContext,
-  cycle: readonly ImportedStaticCssEvalLocalReferenceFrame[]
-): StaticCssEvalDiagnostic {
-  const importChain = cycle.map(formatImportedLocalAliasCycleFrame);
-
-  return createStaticCssEvalDiagnostic({
-    id: "STATIC_CSS_EVAL_LOCAL_ALIAS_CYCLE",
-    code: "cycle-detected",
-    reason: "runtime-dynamic-value",
-    detail: `cyclic imported static css alias detected: ${importChain.join(
-      " -> "
-    )}`,
-    ...createImportedLiteralDiagnosticContext(context),
-    importChain
-  });
-}
-
-function formatImportedLocalAliasCycleFrame(
-  frame: ImportedStaticCssEvalLocalReferenceFrame
-): string {
-  const memberPath = frame.memberPath.length
-    ? `.${frame.memberPath.join(".")}`
-    : "";
-
-  return `${frame.recordId}#${frame.bindingName}${memberPath}`;
-}
-
-function createUnsupportedImportedReferenceResult(
-  options: ImportedStaticCssEvalLiteralReferenceOptions
-): ImportedStaticCssEvalLiteralResult {
-  if (options.reference.kind !== "unsupported") {
-    return createUnsupportedImportedLiteralResult(options, options.expression);
-  }
-
-  if (options.reference.reason === "optional-member-path") {
-    return {
-      kind: "error",
-      diagnostic: createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-        createImportedLiteralDiagnosticContext(options.context),
-        options.reference.detail
-      )
-    };
-  }
-
-  return {
-    kind: "error",
-    diagnostic: createStaticCssEvalComputedMemberUnsupportedDiagnostic(
-      createImportedLiteralDiagnosticContext(options.context)
-    )
-  };
-}
-
-function createDynamicRequireOperandDiagnostic(
-  options: ImportedStaticCssEvalLiteralReferenceOptions
-): StaticCssEvalDiagnostic | null {
-  const binding = options.record.programPath.scope.getBinding(
-    options.reference.bindingName
-  );
-  const init = binding
-    ? getImportedStaticCssEvalVariableBindingInitExpression(binding)
-    : null;
-  const unwrappedInit = init ? unwrapTransparentCssRuleExpression(init) : null;
-
-  if (
-    !unwrappedInit ||
-    !containsStaticCssEvalRequireCallExpression(
-      unwrappedInit,
-      options.record.programPath.scope
-    )
-  ) {
-    return null;
-  }
-
-  const context = createImportedLiteralDiagnosticContext(options.context);
-
-  return isStaticCssEvalLiteralRequireCallExpression(
-    unwrappedInit,
-    options.record.programPath.scope
-  )
-    ? createStaticCssEvalCjsUnsupportedDiagnostic(context)
-    : createStaticCssEvalCjsDynamicRequireUnsupportedDiagnostic(context);
-}
-
-function createUnsupportedImportedLiteralResult(
-  options: Pick<ImportedStaticCssEvalLiteralEvaluationOptions, "context">,
-  expression: t.Expression
-): ImportedStaticCssEvalLiteralResult {
-  const unwrappedExpression = unwrapTransparentCssRuleExpression(expression);
-
-  return {
-    kind: "error",
-    diagnostic: createImportedStaticCssEvalDiagnostic({
-      ...options.context,
-      code: "unsupported-syntax",
-      reason: getUnsupportedLiteralReason(unwrappedExpression),
-      detail: createUnsupportedLiteralDetail(
-        options.context.exportName,
-        unwrappedExpression
-      )
-    })
-  };
-}
-
-function createImportedObjectSpreadError(
-  context: ImportedStaticCssEvalContext,
-  collection: "object" | "array"
-): ImportedStaticCssEvalLiteralResult {
-  return {
-    kind: "error",
-    diagnostic: createStaticCssEvalObjectSpreadUnsupportedDiagnostic(
-      createImportedLiteralDiagnosticContext(context),
-      formatExportName(context.exportName),
-      collection
-    )
-  };
-}
-
-function createImportedLiteralDiagnosticContext(
-  context: ImportedStaticCssEvalContext
-): {
-  readonly owner: StaticCssEvalSourceLocation;
-  readonly dependency: StaticCssEvalSourceLocation;
-  readonly importPath: string;
-  readonly exportName: StaticCssEvalExportName;
-  readonly memberPath: readonly string[];
-} {
-  return {
-    owner: context.owner,
-    dependency: context.dependency,
-    importPath: context.importPath,
-    exportName: context.exportName,
-    memberPath: context.memberPath
-  };
-}
-
-function createReferenceQuery(
-  importerId: string,
-  reference: Extract<StaticMemberReference, { kind: "supported" }>
-): StaticCssEvalQuery {
-  return {
-    importerId,
-    expressionStart: 0,
-    expressionEnd: reference.bindingName.length,
-    bindingName: reference.bindingName,
-    ...(reference.memberPath.length > 0
-      ? { memberPath: [...reference.memberPath] }
-      : {})
-  };
-}
-
-function getImportedStaticCssEvalBindingDeclarationKind(
-  binding: Binding
-): ImportedStaticCssEvalDeclarationKind | null {
-  const bindingPath = binding.path;
-
-  if (bindingPath.isVariableDeclarator()) {
-    if (!t.isIdentifier(bindingPath.node.id)) {
-      return null;
-    }
-
-    if (!bindingPath.parentPath.isVariableDeclaration()) {
-      return null;
-    }
-
-    const { kind } = bindingPath.parentPath.node;
-    return kind === "const" || kind === "let" || kind === "var" ? kind : null;
-  }
-
-  if (bindingPath.isFunctionDeclaration()) {
-    return "function";
-  }
-
-  if (bindingPath.isClassDeclaration()) {
-    return "class";
-  }
-
-  return null;
-}
-
-function getImportedStaticCssEvalVariableBindingInitExpression(
-  binding: Binding
-): t.Expression | null {
-  const bindingPath = binding.path;
-
-  if (
-    !bindingPath.isVariableDeclarator() ||
-    !t.isIdentifier(bindingPath.node.id)
-  ) {
-    return null;
-  }
-
-  const initPath = bindingPath.get("init");
-
-  if (!initPath.node || !initPath.isExpression()) {
-    return null;
-  }
-
-  return initPath.node;
-}
-
-function canUseImportedStaticMemberReferenceFastPath(
-  expression: t.MemberExpression
-): boolean {
-  if (t.isSuper(expression.object)) {
-    return false;
-  }
-
-  if (!isDirectStaticMemberPropertyName(expression)) {
-    return false;
-  }
-
-  const object = unwrapTransparentCssRuleExpression(expression.object);
-
-  if (t.isIdentifier(object)) {
-    return true;
-  }
-
-  return t.isMemberExpression(object)
-    ? canUseImportedStaticMemberReferenceFastPath(object)
-    : false;
-}
-
-function isDirectStaticMemberPropertyName(
-  expression: t.MemberExpression
-): boolean {
-  if (!expression.computed) {
-    return t.isIdentifier(expression.property);
-  }
-
-  return (
-    t.isStringLiteral(expression.property) ||
-    t.isNumericLiteral(expression.property)
-  );
-}
-
-function evaluateStaticCssLiteralExpression(
-  options: ImportedStaticCssEvalLiteralEvaluationOptions
-): ImportedStaticCssEvalLiteralResult {
-  const unwrappedExpression = unwrapTransparentCssRuleExpression(
-    options.expression
-  );
-  options.literalState.count += 1;
-
-  const countResult = enforceImportedStaticCssEvalLiteralCount(
-    options.context,
-    options.literalState
-  );
-
-  if (countResult) {
-    return { kind: "error", diagnostic: countResult };
-  }
-
-  if (t.isObjectExpression(unwrappedExpression)) {
-    const depthResult = enforceImportedStaticCssEvalDepth(
-      options.context,
-      options.depth
-    );
-
-    if (depthResult) {
-      return { kind: "error", diagnostic: depthResult };
-    }
-
-    return evaluateStaticCssObjectExpression({
-      ...options,
-      expression: unwrappedExpression
-    });
-  }
-
-  if (t.isArrayExpression(unwrappedExpression)) {
-    const depthResult = enforceImportedStaticCssEvalDepth(
-      options.context,
-      options.depth
-    );
-
-    if (depthResult) {
-      return { kind: "error", diagnostic: depthResult };
-    }
-
-    return evaluateStaticCssArrayExpression({
-      ...options,
-      expression: unwrappedExpression
-    });
-  }
-
-  if (t.isStringLiteral(unwrappedExpression)) {
-    return { kind: "resolved", value: unwrappedExpression.value };
-  }
-
-  if (t.isNumericLiteral(unwrappedExpression)) {
-    return { kind: "resolved", value: unwrappedExpression.value };
-  }
-
-  if (t.isBooleanLiteral(unwrappedExpression)) {
-    return { kind: "resolved", value: unwrappedExpression.value };
-  }
-
-  if (t.isNullLiteral(unwrappedExpression)) {
-    return { kind: "resolved", value: null };
-  }
-
-  if (
-    t.isUnaryExpression(unwrappedExpression) &&
-    (unwrappedExpression.operator === "+" ||
-      unwrappedExpression.operator === "-") &&
-    t.isNumericLiteral(unwrappedExpression.argument)
-  ) {
-    return {
-      kind: "resolved",
-      value:
-        unwrappedExpression.operator === "-"
-          ? -unwrappedExpression.argument.value
-          : unwrappedExpression.argument.value
-    };
-  }
-
-  if (
-    t.isTemplateLiteral(unwrappedExpression) &&
-    unwrappedExpression.expressions.length === 0
-  ) {
-    const [quasi] = unwrappedExpression.quasis;
-    return {
-      kind: "resolved",
-      value: quasi?.value.cooked ?? quasi?.value.raw ?? ""
-    };
-  }
-
-  if (t.isTemplateLiteral(unwrappedExpression)) {
-    return evaluateStaticCssTemplateLiteralExpression({
-      ...options,
-      expression: unwrappedExpression
-    });
-  }
-
-  if (
-    t.isMemberExpression(unwrappedExpression) ||
-    t.isOptionalMemberExpression(unwrappedExpression)
-  ) {
-    if (
-      t.isMemberExpression(unwrappedExpression) &&
-      canUseImportedStaticMemberReferenceFastPath(unwrappedExpression)
-    ) {
-      const reference = getStaticCssEvalMemberReference(unwrappedExpression);
-
-      if (reference) {
-        return options.resolveReference({
-          ...options,
-          expression: unwrappedExpression,
-          reference
-        });
-      }
-    }
-
-    return evaluateStaticCssMemberExpression({
-      ...options,
-      expression: unwrappedExpression
-    });
-  }
-
-  const reference = getStaticCssEvalMemberReference(unwrappedExpression);
-
-  if (reference) {
-    return options.resolveReference({
-      ...options,
-      expression: unwrappedExpression,
-      reference
-    });
-  }
-
-  return createUnsupportedImportedLiteralResult(options, unwrappedExpression);
-}
-
-function evaluateStaticCssTemplateLiteralExpression(
-  options: ImportedStaticCssEvalLiteralEvaluationOptions & {
-    readonly expression: t.TemplateLiteral;
-  }
-): ImportedStaticCssEvalLiteralResult {
-  let value = "";
-
-  for (let index = 0; index < options.expression.quasis.length; index += 1) {
-    const quasi = options.expression.quasis[index];
-
-    if (quasi) {
-      value += quasi.value.cooked ?? quasi.value.raw;
-    }
-
-    const interpolation = options.expression.expressions[index];
-
-    if (!interpolation) {
-      continue;
-    }
-
-    if (!t.isExpression(interpolation)) {
-      return {
-        kind: "error",
-        diagnostic: createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-          createImportedLiteralDiagnosticContext(options.context),
-          interpolation.type
-        )
-      };
-    }
-
-    const interpolationResult = evaluateStaticCssLiteralExpression({
-      ...options,
-      expression: interpolation,
-      depth: options.depth + 1
-    });
-
-    if (interpolationResult.kind === "error") {
-      return interpolationResult;
-    }
-
-    const primitiveValue = getTemplateInterpolationPrimitiveValue(
-      interpolationResult.value
-    );
-
-    if (primitiveValue === undefined) {
-      return {
-        kind: "error",
-        diagnostic: createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-          createImportedLiteralDiagnosticContext(options.context),
-          interpolation.type
-        )
-      };
-    }
-
-    value += String(primitiveValue);
-  }
-
-  return { kind: "resolved", value };
-}
-
-function evaluateStaticCssMemberExpression(
-  options: ImportedStaticCssEvalLiteralEvaluationOptions & {
-    readonly expression: t.MemberExpression | t.OptionalMemberExpression;
-  }
-): ImportedStaticCssEvalLiteralResult {
-  if (t.isSuper(options.expression.object)) {
-    return {
-      kind: "error",
-      diagnostic: createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-        createImportedLiteralDiagnosticContext(options.context),
-        options.expression.type
-      )
-    };
-  }
-
-  const objectResult = evaluateStaticCssLiteralExpression({
-    ...options,
-    expression: options.expression.object,
-    depth: options.depth + 1
-  });
-
-  if (objectResult.kind === "error") {
-    return objectResult;
-  }
-
-  const memberNameResult = resolveImportedStaticMemberName({
-    ...options,
-    depth: options.depth + 1
-  });
-
-  if (memberNameResult.kind === "error") {
-    return memberNameResult;
-  }
-
-  const memberValue = selectStaticCssLiteralMemberValue(
-    objectResult.value,
-    memberNameResult.memberName
-  );
-
-  return memberValue.kind === "resolved"
-    ? memberValue
-    : {
-        kind: "error",
-        diagnostic: createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-          createImportedLiteralDiagnosticContext(options.context),
-          options.expression.type
-        )
-      };
-}
-
-function evaluateStaticCssObjectExpression(
-  options: ImportedStaticCssEvalLiteralEvaluationOptions & {
-    readonly expression: t.ObjectExpression;
-  }
-): ImportedStaticCssEvalLiteralResult {
-  const value: Record<string, StaticCssLiteral> = Object.create(null);
-
-  for (const property of options.expression.properties) {
-    if (t.isSpreadElement(property)) {
-      const spreadResult = evaluateStaticCssLiteralExpression({
-        ...options,
-        expression: property.argument,
-        depth: options.depth + 1
-      });
-
-      if (spreadResult.kind === "error") {
-        return spreadResult;
-      }
-
-      if (!isStaticCssLiteralObject(spreadResult.value)) {
-        return createImportedObjectSpreadError(options.context, "object");
-      }
-
-      Object.assign(value, spreadResult.value);
-      continue;
-    }
-
-    if (!t.isObjectProperty(property)) {
-      return createStaticCssLiteralError(
-        options.context,
-        "function-or-call",
-        `imported export "${formatExportName(
-          options.context.exportName
-        )}" contains an object method`
-      );
-    }
-
-    if (!t.isExpression(property.value)) {
-      return createStaticCssLiteralError(
-        options.context,
-        "unsupported-literal",
-        `imported export "${formatExportName(
-          options.context.exportName
-        )}" contains a non-expression object value`
-      );
-    }
-
-    const keyResult = resolveImportedStaticObjectPropertyKey({
-      ...options,
-      property,
-      depth: options.depth + 1
-    });
-
-    if (keyResult.kind === "error") {
-      return keyResult;
-    }
-
-    const propertyResult = evaluateStaticCssLiteralExpression({
-      ...options,
-      expression: property.value,
-      depth: options.depth + 1
-    });
-
-    if (propertyResult.kind === "error") {
-      return propertyResult;
-    }
-
-    value[keyResult.propertyName] = propertyResult.value;
-  }
-
-  return { kind: "resolved", value };
-}
-
-function resolveImportedStaticObjectPropertyKey(
-  options: ImportedStaticCssEvalLiteralEvaluationOptions & {
-    readonly property: t.ObjectProperty;
-  }
-):
-  | { kind: "resolved"; propertyName: string }
-  | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-  const staticName = getStaticObjectPropertyName(options.property.key);
-
-  if (!options.property.computed) {
-    return staticName
-      ? { kind: "resolved", propertyName: staticName }
-      : createStaticCssLiteralError(
-          options.context,
-          "unsupported-literal",
-          `imported export "${formatExportName(
-            options.context.exportName
-          )}" contains an unsupported object key`
-        );
-  }
-
-  if (!t.isExpression(options.property.key)) {
-    return {
-      kind: "error",
-      diagnostic: createStaticCssEvalComputedMemberUnsupportedDiagnostic(
-        createImportedLiteralDiagnosticContext(options.context)
-      )
-    };
-  }
-
-  const keyResult = evaluateStaticCssLiteralExpression({
-    ...options,
-    expression: options.property.key
-  });
-
-  if (keyResult.kind === "error") {
-    return keyResult.diagnostic.id
-      ? keyResult
-      : {
-          kind: "error",
-          diagnostic: createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-            createImportedLiteralDiagnosticContext(options.context),
-            options.property.key.type
-          )
-        };
-  }
-
-  const propertyName = getStaticStringOrNumberLiteralValue(keyResult.value);
-
-  return propertyName === null
-    ? {
-        kind: "error",
-        diagnostic: createStaticCssEvalComputedMemberUnsupportedDiagnostic(
-          createImportedLiteralDiagnosticContext(options.context)
-        )
-      }
-    : { kind: "resolved", propertyName: String(propertyName) };
-}
-
-function resolveImportedStaticMemberName(
-  options: ImportedStaticCssEvalLiteralEvaluationOptions & {
-    readonly expression: t.MemberExpression | t.OptionalMemberExpression;
-  }
-):
-  | { kind: "resolved"; memberName: string }
-  | { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-  if (
-    !options.expression.computed &&
-    t.isIdentifier(options.expression.property)
-  ) {
-    return { kind: "resolved", memberName: options.expression.property.name };
-  }
-
-  if (!t.isExpression(options.expression.property)) {
-    return {
-      kind: "error",
-      diagnostic: createStaticCssEvalComputedMemberUnsupportedDiagnostic(
-        createImportedLiteralDiagnosticContext(options.context)
-      )
-    };
-  }
-
-  const keyResult = evaluateStaticCssLiteralExpression({
-    ...options,
-    expression: options.expression.property
-  });
-
-  if (keyResult.kind === "error") {
-    return keyResult.diagnostic.id
-      ? keyResult
-      : {
-          kind: "error",
-          diagnostic: createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-            createImportedLiteralDiagnosticContext(options.context),
-            options.expression.property.type
-          )
-        };
-  }
-
-  const memberName = getStaticStringOrNumberLiteralValue(keyResult.value);
-
-  return memberName === null
-    ? {
-        kind: "error",
-        diagnostic: createStaticCssEvalComputedMemberUnsupportedDiagnostic(
-          createImportedLiteralDiagnosticContext(options.context)
-        )
-      }
-    : { kind: "resolved", memberName: String(memberName) };
-}
-
-function evaluateStaticCssArrayExpression(
-  options: ImportedStaticCssEvalLiteralEvaluationOptions & {
-    readonly expression: t.ArrayExpression;
-  }
-): ImportedStaticCssEvalLiteralResult {
-  const value: StaticCssLiteral[] = [];
-
-  for (const element of options.expression.elements) {
-    if (!element) {
-      return createStaticCssLiteralError(
-        options.context,
-        "unsupported-literal",
-        `imported export "${formatExportName(
-          options.context.exportName
-        )}" contains an array hole`
-      );
-    }
-
-    if (t.isSpreadElement(element)) {
-      const spreadResult = evaluateStaticCssLiteralExpression({
-        ...options,
-        expression: element.argument,
-        depth: options.depth + 1
-      });
-
-      if (spreadResult.kind === "error") {
-        return spreadResult;
-      }
-
-      if (!Array.isArray(spreadResult.value)) {
-        return createImportedObjectSpreadError(options.context, "array");
-      }
-
-      value.push(...spreadResult.value);
-      continue;
-    }
-
-    const elementResult = evaluateStaticCssLiteralExpression({
-      ...options,
-      expression: element,
-      depth: options.depth + 1
-    });
-
-    if (elementResult.kind === "error") {
-      return elementResult;
-    }
-
-    value.push(elementResult.value);
-  }
-
-  return { kind: "resolved", value };
-}
-
-function selectStaticCssLiteralMemberValue(
-  value: StaticCssLiteral,
-  memberName: string
-): { kind: "resolved"; value: StaticCssLiteral } | { kind: "not-candidate" } {
-  if (isStaticCssLiteralObject(value)) {
-    if (!Object.prototype.hasOwnProperty.call(value, memberName)) {
-      return { kind: "not-candidate" };
-    }
-
-    const memberValue = value[memberName];
-
-    return memberValue === undefined
-      ? { kind: "not-candidate" }
-      : { kind: "resolved", value: memberValue };
-  }
-
-  if (Array.isArray(value) && /^\d+$/.test(memberName)) {
-    const memberValue = value[Number(memberName)];
-
-    return memberValue === undefined
-      ? { kind: "not-candidate" }
-      : { kind: "resolved", value: memberValue };
-  }
-
-  return { kind: "not-candidate" };
-}
-
-function findUnsupportedImportedObjectReferenceDiagnostic(
-  expression: t.ObjectExpression,
-  options: {
-    ownerFile: string;
-    provider: StaticCssEvalProvider;
-    includeSupportedReferenceErrors?: boolean;
-  }
-): StaticCssEvalDiagnostic | null {
-  for (const property of expression.properties) {
-    if (!t.isObjectProperty(property) || !t.isExpression(property.value)) {
-      continue;
-    }
-
-    const diagnostic = findUnsupportedImportedExpressionReferenceDiagnostic(
-      property.value,
-      options
-    );
-
-    if (diagnostic) {
-      return diagnostic;
-    }
-  }
-
-  return null;
-}
-
-function findUnsupportedImportedArrayReferenceDiagnostic(
-  expression: t.ArrayExpression,
-  options: {
-    ownerFile: string;
-    provider: StaticCssEvalProvider;
-    includeSupportedReferenceErrors?: boolean;
-  }
-): StaticCssEvalDiagnostic | null {
-  for (const element of expression.elements) {
-    if (!element || t.isSpreadElement(element)) {
-      continue;
-    }
-
-    const diagnostic = findUnsupportedImportedExpressionReferenceDiagnostic(
-      element,
-      options
-    );
-
-    if (diagnostic) {
-      return diagnostic;
-    }
-  }
-
-  return null;
-}
-
-function findUnsupportedImportedExpressionReferenceDiagnostic(
-  expression: t.Expression,
-  options: {
-    ownerFile: string;
-    provider: StaticCssEvalProvider;
-    includeSupportedReferenceErrors?: boolean;
-  }
-): StaticCssEvalDiagnostic | null {
-  const unwrappedExpression = unwrapTransparentCssRuleExpression(expression);
-  const reference = getStaticCssEvalMemberReference(unwrappedExpression);
-
-  if (reference?.kind === "unsupported") {
-    const result = options.provider.getResolvedCssValue({
-      importerId: options.ownerFile,
-      expressionStart: unwrappedExpression.start ?? 0,
-      expressionEnd: unwrappedExpression.end ?? 0,
-      bindingName: reference.bindingName,
-      ...(reference.memberPath.length > 0
-        ? { memberPath: reference.memberPath }
-        : {})
-    });
-
-    if (result.kind === "not-candidate") {
-      return null;
-    }
-
-    if (reference.reason === "dynamic-member-path") {
-      return createStaticCssEvalComputedMemberUnsupportedDiagnostic({
-        owner: createExpressionOwnerLocation(
-          unwrappedExpression,
-          options.ownerFile
-        ),
-        memberPath: reference.memberPath
-      });
-    }
-
-    return result.kind === "error"
-      ? result.diagnostic
-      : createStaticCssEvalDynamicExpressionUnsupportedDiagnostic(
-          {
-            owner: createExpressionOwnerLocation(
-              unwrappedExpression,
-              options.ownerFile
-            )
-          },
-          reference.detail
-        );
-  }
-
-  const candidate = createStaticCssEvalCandidate(
-    unwrappedExpression,
-    options.ownerFile
-  );
-
-  if (
-    !candidate &&
-    t.isMemberExpression(unwrappedExpression) &&
-    unwrappedExpression.computed &&
-    t.isIdentifier(unwrappedExpression.property)
-  ) {
-    return createStaticCssEvalComputedMemberUnsupportedDiagnostic({
-      owner: createExpressionOwnerLocation(
-        unwrappedExpression,
-        options.ownerFile
-      )
-    });
-  }
-
-  if (
-    options.includeSupportedReferenceErrors !== false &&
-    candidate?.bindingName
-  ) {
-    const result = options.provider.getResolvedCssValue(candidate);
-
-    if (result.kind === "error") {
-      return result.diagnostic;
-    }
-  }
-
-  if (t.isObjectExpression(unwrappedExpression)) {
-    return findUnsupportedImportedObjectReferenceDiagnostic(
-      unwrappedExpression,
-      options
-    );
-  }
-
-  if (t.isArrayExpression(unwrappedExpression)) {
-    return findUnsupportedImportedArrayReferenceDiagnostic(
-      unwrappedExpression,
-      options
-    );
-  }
-
-  return null;
-}
-
-function enforceImportedStaticCssEvalDepth(
-  context: ImportedStaticCssEvalContext,
-  depth: number
-): StaticCssEvalDiagnostic | null {
-  const result = enforceStaticCssEvalObjectArrayRecursionDepth({
-    owner: context.owner,
-    dependency: context.dependency,
-    importPath: context.importPath,
-    exportName: context.exportName,
-    memberPath: context.memberPath,
-    recursionDepth: depth
-  });
-
-  return result.ok ? null : result.diagnostic;
-}
-
-function enforceImportedStaticCssEvalLiteralCount(
-  context: ImportedStaticCssEvalContext,
-  state: StaticCssLiteralValidationState
-): StaticCssEvalDiagnostic | null {
-  const result = enforceStaticCssEvalLiteralNodeCount({
-    owner: context.owner,
-    dependency: context.dependency,
-    importPath: context.importPath,
-    exportName: context.exportName,
-    memberPath: context.memberPath,
-    literalNodeCount: state.count
-  });
-
-  return result.ok ? null : result.diagnostic;
-}
-
-function createStaticCssLiteralError(
-  context: ImportedStaticCssEvalContext,
-  reason: StaticCssEvalUnsupportedReason,
-  detail: string
-): { kind: "error"; diagnostic: StaticCssEvalDiagnostic } {
-  return {
-    kind: "error",
-    diagnostic: createImportedStaticCssEvalDiagnostic({
-      ...context,
-      code: "unsupported-syntax",
-      reason,
-      detail
-    })
-  };
-}
-
-function createImportedStaticCssEvalDiagnostic(
-  options: ImportedStaticCssEvalContext & {
-    code: StaticCssEvalDiagnostic["code"];
-    reason: StaticCssEvalUnsupportedReason;
-    detail: string;
-  }
-): StaticCssEvalDiagnostic {
-  return createStaticCssEvalDiagnostic({
-    code: options.code,
-    reason: options.reason,
-    detail: options.detail,
-    owner: options.owner,
-    dependency: options.dependency,
-    importPath: options.importPath,
-    exportName: options.exportName,
-    memberPath: options.memberPath,
-    importChain: [
-      options.owner.file,
-      `${options.dependency.file}#${formatExportName(options.exportName)}${
-        options.memberPath.length > 0 ? `.${options.memberPath.join(".")}` : ""
-      }`
-    ]
-  });
-}
-
-function createQueryOwnerLocation(
-  query: StaticCssEvalQuery
-): StaticCssEvalSourceLocation {
-  return {
-    file: query.importerId,
-    start: query.expressionStart,
-    end: query.expressionEnd
-  };
-}
-
-function createExpressionOwnerLocation(
-  expression: t.Expression,
-  ownerFile: string
-): StaticCssEvalSourceLocation {
-  return {
-    file: ownerFile,
-    ...(typeof expression.start === "number"
-      ? { start: expression.start }
-      : {}),
-    ...(typeof expression.end === "number" ? { end: expression.end } : {})
-  };
-}
-
-function getModuleStringName(
-  node: t.Identifier | t.StringLiteral
-): string | null {
-  if (t.isIdentifier(node)) {
-    return node.name;
-  }
-
-  if (t.isStringLiteral(node)) {
-    return node.value;
-  }
-
-  return null;
-}
-
-function getStaticStringOrNumberLiteralValue(
-  value: StaticCssLiteral
-): string | number | null {
-  return typeof value === "string" || typeof value === "number" ? value : null;
-}
-
-function getTemplateInterpolationPrimitiveValue(
-  value: StaticCssLiteral
-): string | number | boolean | null | undefined {
-  if (
-    typeof value === "string" ||
-    typeof value === "number" ||
-    typeof value === "boolean" ||
-    value === null
-  ) {
-    return value;
-  }
-
-  return undefined;
-}
-function createUnsupportedLiteralDetail(
-  exportName: StaticCssEvalExportName,
-  expression: t.Expression
-): string {
-  if (getUnsupportedLiteralReason(expression) === "dynamic-import") {
-    return `imported export "${formatExportName(exportName)}" contains a dynamic import`;
-  }
-
-  return `imported export "${formatExportName(
-    exportName
-  )}" contains unsupported ${expression.type}`;
-}
-function isStaticCssRuleLiteralValue(
-  value: StaticCssLiteral
-): value is StaticCssLiteral[] | { [key: string]: StaticCssLiteral } {
-  return value !== null && typeof value === "object";
-}
-
-function isStaticCssLiteralObject(
-  value: StaticCssLiteral
-): value is { [key: string]: StaticCssLiteral } {
-  return value !== null && typeof value === "object" && !Array.isArray(value);
-}
-
-function formatExportName(exportName: StaticCssEvalExportName): string {
-  return exportName === null ? "<local>" : exportName;
-}
-
-function createImportResolutionKey(
-  importerId: string,
-  importPath: string
-): string {
-  return `${importerId}${importResolutionKeySeparator}${importPath}`;
-}
+import type {
+  ImportedStaticCssEvalImportResolution,
+  ImportedStaticCssEvalLoadedModule,
+  ImportedStaticCssEvalResolutionState
+} from "./importedModules/contracts.js";
+import { createReferenceQuery } from "./importedModules/localReference.js";
+import { createImportedStaticCssEvalModuleRecord } from "./importedModules/provider.js";
+import {
+  addResolutionDependency,
+  getResolutionDependencies
+} from "./importedModules/resolution.js";
+import { findUnsupportedImportedStaticCssEvalReferenceDiagnostic } from "./importedModules/expression.js";
+import { ImportedStaticCssEvalResolver } from "./importedModules/resolver.js";
+
+export type {
+  CreateImportedStaticCssEvalProviderOptions,
+  ImportedStaticCssEvalCjsBinding,
+  ImportedStaticCssEvalExportBinding,
+  ImportedStaticCssEvalImportBinding,
+  ImportedStaticCssEvalImportResolution,
+  ImportedStaticCssEvalLoadedModule,
+  ImportedStaticCssEvalModuleRecord,
+  ResolveImportedStaticCssEvalExpressionResult
+} from "./importedModules/contracts.js";
+export {
+  createImportedStaticCssEvalModuleRecord,
+  createImportedStaticCssEvalProvider
+} from "./importedModules/provider.js";
+export {
+  createStaticCssLiteralExpression,
+  findUnsupportedImportedStaticCssEvalReferenceDiagnostic,
+  resolveImportedStaticCssEvalExpression
+} from "./importedModules/expression.js";
 
 // == Tests ====================================================================
+
 // Ignore errors when compiling to CommonJS.
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore error TS1343: The 'import.meta' meta-property is only allowed when the '--module' option is 'es2020', 'es2022', 'esnext', 'system', 'node16', or 'nodenext'.
@@ -4139,6 +243,71 @@ if (import.meta.vitest) {
   }
 
   describe("imported static css eval module records", () => {
+    it("preserves reference ranges in nested resolution queries", () => {
+      const reference = {
+        kind: "supported" as const,
+        bindingName: "styles",
+        memberPath: ["button"],
+        expressionStart: 42,
+        expressionEnd: 57
+      };
+
+      expect(createReferenceQuery(ownerId, reference)).toEqual({
+        importerId: ownerId,
+        expressionStart: 42,
+        expressionEnd: 57,
+        bindingName: "styles",
+        memberPath: ["button"]
+      });
+    });
+
+    it("preserves watch files when resolution dependencies merge", () => {
+      const state: ImportedStaticCssEvalResolutionState = {
+        owner: { file: ownerId },
+        dependencies: new Map(),
+        resolutionChain: [],
+        stack: []
+      };
+      const dependency = {
+        file: stylesId,
+        kind: "imported",
+        importer: ownerId,
+        specifier: "./styles",
+        exportName: "button",
+        memberPath: [],
+        inspected: true,
+        contributed: true,
+        sourceKind: "project-source",
+        sourceOrigin: "project",
+        canonicalModuleId: "project:styles",
+        normalizedPathKey: "/project/src/styles.ts",
+        watchFiles: ["/project/.pnp.cjs"],
+        unsupportedReason: "unresolved"
+      } satisfies ResolutionDependency;
+
+      addResolutionDependency(state, dependency);
+      addResolutionDependency(state, {
+        ...dependency,
+        sourceKind: undefined,
+        sourceOrigin: undefined,
+        canonicalModuleId: undefined,
+        normalizedPathKey: undefined,
+        watchFiles: undefined,
+        unsupportedReason: undefined
+      });
+
+      expect(getResolutionDependencies(state)).toEqual([
+        expect.objectContaining({
+          sourceKind: "project-source",
+          sourceOrigin: "project",
+          canonicalModuleId: "project:styles",
+          normalizedPathKey: "/project/src/styles.ts",
+          watchFiles: ["/project/.pnp.cjs"],
+          unsupportedReason: "unresolved"
+        })
+      ]);
+    });
+
     it("matches parser plugins to the loaded module extension", () => {
       expect(() =>
         createImportedStaticCssEvalModuleRecord({
@@ -4175,6 +344,59 @@ if (import.meta.vitest) {
       expect(red.sourceHash).toMatch(/^inline:[a-z0-9]+$/);
       expect(red.sourceHash).toBe(red.parsedModule.sourceHash);
       expect(red.sourceHash).not.toBe(sky.sourceHash);
+    });
+
+    it("reports imported module parse failures in unresolved diagnostics", () => {
+      expect(
+        resolveFixture(createProvider(`export const button = {`), "button")
+      ).toMatchObject({
+        kind: "error",
+        diagnostic: {
+          id: "STATIC_CSS_EVAL_UNRESOLVED_IMPORT",
+          message: expect.stringMatching(
+            /Unexpected token|Unexpected end of input/
+          )
+        }
+      });
+    });
+
+    it("preserves non-Error imported module parse failures", () => {
+      const parseFailure = "synthetic parse failure";
+      const delegate = createStaticCssModuleCache();
+      const moduleCache: StaticCssModuleCache = {
+        ...delegate,
+        getParsedModule(source) {
+          if (source.resolvedFile === stylesId) {
+            throw parseFailure;
+          }
+
+          return delegate.getParsedModule(source);
+        }
+      };
+      const provider = createProviderFromModules({
+        modules: [
+          {
+            id: ownerId,
+            source: `import { button } from "./styles"; <div css={button} />;`
+          },
+          {
+            id: stylesId,
+            source: `export const button = { color: "red" } as const;`
+          }
+        ],
+        importResolutions: [
+          { importerId: ownerId, importPath: "./styles", resolvedId: stylesId }
+        ],
+        moduleCache
+      });
+
+      expect(resolveFixture(provider, "button")).toMatchObject({
+        kind: "error",
+        diagnostic: {
+          id: "STATIC_CSS_EVAL_UNRESOLVED_IMPORT",
+          message: expect.stringContaining(parseFailure)
+        }
+      });
     });
 
     it("collects literal CommonJS require bindings into module records", () => {
@@ -4571,6 +793,23 @@ if (import.meta.vitest) {
           })
         ])
       );
+    });
+
+    it("rejects non-canonical imported array member names", () => {
+      const result = resolveFixture(
+        createProvider(
+          `export const button = { color: ["red", "blue"]["01"] } as const;`,
+          `import { button } from "./styles"; <div css={button} />;`
+        ),
+        "button"
+      );
+
+      expect(result).toMatchObject({
+        kind: "error",
+        diagnostic: {
+          id: "STATIC_CSS_EVAL_DYNAMIC_EXPRESSION_UNSUPPORTED"
+        }
+      });
     });
 
     it("resolves namespace export-star computed optional imported static css eval operands", () => {
@@ -5224,33 +1463,6 @@ if (import.meta.vitest) {
           exportName: "dynamic"
         }
       });
-    });
-
-    it("preserves special property names on whole namespace values", () => {
-      const result = expectResolvedResult(
-        resolveFixture(
-          createProvider(
-            `
-              const proto = { color: "red" } as const;
-              const constructorValue = { color: "blue" } as const;
-              export { proto as "__proto__", constructorValue as "constructor" };
-            `,
-            `import * as styles from "./styles"; <div css={styles} />;`
-          ),
-          "styles"
-        )
-      );
-
-      if (
-        result.value === null ||
-        typeof result.value !== "object" ||
-        Array.isArray(result.value)
-      ) {
-        throw new TypeError("expected a static namespace object");
-      }
-
-      expect(Object.hasOwn(result.value, "__proto__")).toBe(true);
-      expect(Object.hasOwn(result.value, "constructor")).toBe(true);
     });
 
     it("preserves special property names on imported object literals", () => {
@@ -6160,59 +2372,6 @@ if (import.meta.vitest) {
       });
     });
 
-    it("reports imported module parse failures in unresolved diagnostics", () => {
-      expect(
-        resolveFixture(createProvider(`export const button = {`), "button")
-      ).toMatchObject({
-        kind: "error",
-        diagnostic: {
-          id: "STATIC_CSS_EVAL_UNRESOLVED_IMPORT",
-          message: expect.stringMatching(
-            /Unexpected token|Unexpected end of input/
-          )
-        }
-      });
-    });
-
-    it("preserves non-Error imported module parse failures", () => {
-      const parseFailure = "synthetic parse failure";
-      const delegate = createStaticCssModuleCache();
-      const moduleCache: StaticCssModuleCache = {
-        ...delegate,
-        getParsedModule(source) {
-          if (source.resolvedFile === stylesId) {
-            throw parseFailure;
-          }
-
-          return delegate.getParsedModule(source);
-        }
-      };
-      const provider = createProviderFromModules({
-        modules: [
-          {
-            id: ownerId,
-            source: `import { button } from "./styles"; <div css={button} />;`
-          },
-          {
-            id: stylesId,
-            source: `export const button = { color: "red" } as const;`
-          }
-        ],
-        importResolutions: [
-          { importerId: ownerId, importPath: "./styles", resolvedId: stylesId }
-        ],
-        moduleCache
-      });
-
-      expect(resolveFixture(provider, "button")).toMatchObject({
-        kind: "error",
-        diagnostic: {
-          id: "STATIC_CSS_EVAL_UNRESOLVED_IMPORT",
-          message: expect.stringContaining(parseFailure)
-        }
-      });
-    });
-
     it("resolves static CommonJS require bindings and require-backed CJS exports", () => {
       const cjsImportProvider = createProvider(
         `
@@ -6978,6 +3137,57 @@ if (import.meta.vitest) {
         id: "STATIC_CSS_EVAL_COMPUTED_MEMBER_UNSUPPORTED"
       });
 
+      const optionalDiagnostic =
+        findUnsupportedImportedStaticCssEvalReferenceDiagnostic({
+          expression: t.optionalMemberExpression(
+            t.identifier("styles"),
+            t.identifier("button"),
+            false,
+            true
+          ),
+          ownerFile: ownerId,
+          provider
+        });
+
+      expect(optionalDiagnostic).toMatchObject({
+        id: "STATIC_CSS_EVAL_DYNAMIC_EXPRESSION_UNSUPPORTED",
+        expressionType: "OptionalMemberExpression"
+      });
+      expect(optionalDiagnostic?.message).toContain(
+        "optional member paths are unsupported"
+      );
+
+      for (const expression of [
+        t.objectExpression([
+          t.spreadElement(
+            t.memberExpression(
+              t.identifier("styles"),
+              t.identifier("variant"),
+              true
+            )
+          )
+        ]),
+        t.arrayExpression([
+          t.spreadElement(
+            t.memberExpression(
+              t.identifier("styles"),
+              t.identifier("variant"),
+              true
+            )
+          )
+        ])
+      ]) {
+        expect(
+          findUnsupportedImportedStaticCssEvalReferenceDiagnostic({
+            expression,
+            ownerFile: ownerId,
+            provider
+          })
+        ).toMatchObject({
+          id: "STATIC_CSS_EVAL_COMPUTED_MEMBER_UNSUPPORTED"
+        });
+      }
+
       expect(
         resolveFixture(
           createProvider(
@@ -7021,6 +3231,46 @@ if (import.meta.vitest) {
           id: "STATIC_CSS_EVAL_DYNAMIC_EXPRESSION_UNSUPPORTED"
         }
       });
+    });
+
+    it("resolves numeric imported array member paths", () => {
+      const result = resolveFixture(
+        createProvider(
+          `export const styles = { cards: [{ color: "red" }] } as const;`,
+          `import { styles } from "./styles"; <div css={styles.cards[0].color} />;`
+        ),
+        "styles",
+        ["cards", "0", "color"]
+      );
+
+      expect(expectResolvedResult(result).value).toBe("red");
+    });
+
+    it("defers imported array member paths across preceding spreads", () => {
+      const result = resolveFixture(
+        createProvider(
+          `const baseCards = [{ color: "green" }, { color: "yellow" }] as const;
+          export const styles = { cards: [{ color: "red" }, ...baseCards, { color: "blue" }] } as const;`,
+          `import { styles } from "./styles"; <div css={styles.cards[2].color} />;`
+        ),
+        "styles",
+        ["cards", "2", "color"]
+      );
+
+      expect(expectResolvedResult(result).value).toBe("yellow");
+    });
+
+    it("preserves holes in imported array member paths", () => {
+      const result = resolveFixture(
+        createProvider(
+          `export const styles = { cards: [{ color: "red" }, , { color: "blue" }] } as const;`,
+          `import { styles } from "./styles"; <div css={styles.cards[2].color} />;`
+        ),
+        "styles",
+        ["cards", "2", "color"]
+      );
+
+      expect(expectResolvedResult(result).value).toBe("blue");
     });
 
     it("rejects unsupported namespace reexports with exact diagnostics", () => {
