+++ packages/babel/src/staticCssEval/importedModules/format.ts
@@ -1,0 +1,5 @@
+import type { StaticCssEvalExportName } from "../types.js";
+
+export function formatExportName(exportName: StaticCssEvalExportName): string {
+  return exportName === null ? "<local>" : exportName;
+}
